% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_parameters.R
\name{change_format.choice_parameters}
\alias{change_format.choice_parameters}
\title{TODO: inherit choice_effects and error_term}
\usage{
\method{change_format}{choice_parameters}(x, new_format = NULL, choice_effects, names = "effect")
}
\arguments{
\item{x}{[\code{choice_parameters} | \code{numeric()}]\cr
The choice_parameters.}

\item{names}{[\code{character(1)}]\cr
\itemize{
\item \code{"effect"}
\item \code{"generic"}
\item \code{"unnamed"}
}

TODO: order of numeric vector}
}
\description{
TODO: inherit choice_effects and error_term
}
