% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_check.R
\name{check_model_consistent}
\alias{check_model_consistent}
\title{Checks consistency of model}
\usage{
check_model_consistent()
}
\arguments{
\item{choice_formula}{TODO}
}
\value{
A \code{list} with the elements
\itemize{
\item \code{ordered}, either \code{TRUE} for ordered choices, or
\code{FALSE} if not, or \code{NA} if not determined,
\item \code{ranked}, either \code{TRUE} for a ranked model, or \code{FALSE}
if not, or \code{NA} if not determined.
}
}
\description{
TODO
}
