% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_response.R
\name{choice_response}
\alias{choice_response}
\alias{simulate_choice_response}
\alias{is.choice_response}
\title{Define choice response}
\usage{
choice_response(choices, choice_identifiers, column_choice = "choice")

simulate_choice_response(
  choice_effects,
  choice_covariates,
  choice_parameters,
  choice_preferences,
  column_choice = "choice"
)

is.choice_response(x, error = TRUE)
}
\arguments{
\item{choices}{TODO}

\item{choice_identifiers}{TODO}

\item{column_choice}{A \code{character}, the column name of the \code{data.frame} with the
choices.}

\item{choice_covariates}{[\code{choice_covariates}]\cr
The choice covariates used for the choice data simulation.}

\item{choice_parameters}{An \code{\link{choice_parameters}} object, which contains the model
parameters used for the choice data simulation.}

\item{choice_preferences}{TODO}

\item{x}{TODO}

\item{error}{TODO}

\item{choice_formula}{TODO}

\item{...}{TODO}
}
\value{
A \code{\link{choice_response}} object. It can be a \code{data.frame} or
\code{list} object and can be transformed between these two formats via
\code{as.data.frame.choices()} or \code{as.list.choices()}:
\itemize{
\item \code{data.frame} format: a \code{data.frame} with three columns
named, \code{column_decider} (the decider identifiers),
\code{column_occasion} (the identifiers for the choice occasions),
and \code{column_choice} (the choices)
\item \code{list} format: a \code{list}, where the \code{n}-th element is
a \code{vector} of choices for the \code{n}-th decider, where the
\code{t}-th element is the choice at their \code{t}-th occasion
}
The \code{\link{choice_response}} object has the following attributes:
\itemize{
\item TODO
}
}
\description{
These functions construct, simulate, and transform an object of class
\code{choice_response}, which defines the observed discrete choices.
\itemize{
\item \code{choice_response()} constructs a \code{choice_response} object.
\item \code{simulate_choice_response()} simulates choices.
\item \code{change_format.choice_parameters()} transforms between \code{list} and
\code{data.frame} format, see details.
}
}
