% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_responses.R
\name{choice_responses}
\alias{choice_responses}
\alias{generate_choice_responses}
\title{Define choice response}
\usage{
choice_responses(
  data_frame,
  column_choice = "choice",
  column_decider = "deciderID",
  column_occasion = NULL,
  cross_section = is.null(column_occasion)
)

generate_choice_responses(
  choice_effects,
  choice_covariates = generate_choice_covariates(choice_effects = choice_effects),
  choice_parameters = generate_choice_parameters(choice_effects = choice_effects),
  choice_identifiers = extract_choice_identifiers(choice_covariates),
  choice_preferences = generate_choice_preferences(choice_parameters = choice_parameters,
    choice_effects = choice_effects, choice_identifiers = choice_identifiers),
  column_choice = "choice",
  choice_type = c("unordered", "ordered", "ranked")
)
}
\arguments{
\item{data_frame}{[\code{data.frame}]\cr
Contains the choice responses.}

\item{column_choice}{[\code{character(1)}]\cr
The column name of \code{data_frame} with the choice responses.}

\item{column_decider}{[\code{character(1)}]\cr
The name of the identifier column for deciders.}

\item{column_occasion}{[\code{character(1)} | \code{NULL}]\cr
The name of the identifier column for choice occasions (panel data).
Can be \code{NULL} for the cross-sectional case.}

\item{cross_section}{[\code{logical(1)}]\cr
Treat choice data as cross-sectional?}

\item{choice_effects}{[\code{choice_effects}]\cr
A \code{\link{choice_effects}} object.}

\item{choice_covariates}{[\code{choice_covariates}]\cr
A \code{\link{choice_covariates}} object.}

\item{choice_parameters}{[\code{choice_parameters}]\cr
A \code{\link{choice_parameters}} object.}

\item{choice_identifiers}{[\code{choice_identifiers}]\cr
A \code{\link{choice_identifiers}} object.}

\item{choice_preferences}{[\code{choice_preferences}]\cr
A \code{\link{choice_preferences}} object.}

\item{choice_type}{[\code{character(1)}]\cr
The response type to simulate. Use \code{"unordered"} (default), \code{"ordered"},
or \code{"ranked"}.}
}
\value{
A \code{choice_responses} tibble.
}
\description{
The \code{choice_responses} object defines the observed choice responses.
\itemize{
\item \code{generate_choice_responses()} simulates choices
}
}
\examples{
### generate choice responses from choice effects
choice_effects <- choice_effects(
  choice_formula = choice_formula(
    formula = choice ~ price | time,
    error_term = "probit"
  ),
  choice_alternatives = choice_alternatives(J = 5)
)
(generate_choice_responses(
  choice_effects = choice_effects,
  choice_type = "ranked"
))
}
\keyword{data}
