% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{choices}
\alias{choices}
\alias{simulate_choices}
\alias{as.data.frame.choices}
\alias{as.list.choices}
\title{Define choices}
\usage{
choices()

simulate_choices(
  choice_parameters,
  choice_covariates,
  choice_set,
  seed = NULL,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc"
)

\method{as.data.frame}{choices}()

\method{as.list}{choices}()
}
\value{
A \code{\link{choices}} object.
}
\description{
These functions construct and validate an object of class
\code{\link{choices}}, which defines the observed choices.
}
