% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{decider_ids_to_cross_section}
\alias{decider_ids_to_cross_section}
\title{Unique decider ids in cross-sectional case}
\usage{
decider_ids_to_cross_section(decider_ids, occasion_ids, delimiter = ".")
}
\arguments{
\item{decider_ids, occasion_ids}{[\code{atomic()}]\cr
An \code{atomic} \code{vector} of ids.}
}
\value{
An \code{atomic} \code{vector} of unique ids.
}
\description{
This helper function makes unique decider ids for a given combination of
decider and occasion ids.
}
\keyword{internal}
