% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{doc-helper}
\alias{doc-helper}
\title{Documentation helper}
\arguments{
\item{choice_alternatives}{A \code{\link{choice_alternatives}} object.}

\item{choice_formula}{A \code{\link{choice_formula}} object.}

\item{error}{Either \code{TRUE} to throw an error if the check failed, or \code{FALSE}
to return the check result as \code{TRUE} or \code{FALSE}, respectively.}

\item{x}{An object of the corresponding class to be printed.}

\item{...}{Currently not used.}
}
\description{
Documentation helper
}
\keyword{internal}
