% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{expand_Tp}
\alias{expand_Tp}
\title{Expand \code{Tp}}
\usage{
expand_Tp(N, Tp = 1)
}
\arguments{
\item{N}{(\code{integer(1)})\cr
The number of deciders.}

\item{Tp}{(\code{integer()})\cr
The number of choice occasions per decider.

Can also be a \code{vector} of length \code{N} for a variable number of
choice occasions per decider.}
}
\value{
An \code{integer} \code{vector} of length \code{N}.
}
\description{
This helper function expands the number of choice occasions \code{Tp} to a
\code{vector} of length \code{N}.
}
\keyword{internal}
