% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fishing_choice}
\alias{fishing_choice}
\title{Recreational Fishing Mode Choice}
\format{
A \code{tibble} with 1182 rows and 11 columns:
\describe{
\item{respondent [\code{integer}]}{The respondent identifier.}
\item{choice [\code{character}]}{The chosen fishing mode.}
\item{price_beach, price_pier, price_boat, price_charter
[\code{numeric}]}{The cost of a fishing trip for each mode, in US dollars.}
\item{catch_beach, catch_pier, catch_boat, catch_charter
[\code{numeric}]}{The expected catch rate for each fishing mode.}
\item{income [\code{numeric}]}{The respondent's monthly income in US dollars.}
}
}
\source{
Adapted from \code{Fishing} in the
\href{https://CRAN.R-project.org/package=mlogit}{\pkg{mlogit} package}.
The original study is described by Herriges and Kling (1999).
}
\usage{
fishing_choice
}
\description{
Choices by 1182 US respondents among fishing options.

This data can be used to model the choice between beach, pier, private-boat,
and charter-boat fishing based on trip cost, expected catch rate, and
monthly income.
}
\references{
\insertRef{HerrigesKling1999}{choicedata}
}
\keyword{dataset}
