% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_parameters.R
\name{get_coefficient_vector}
\alias{get_coefficient_vector}
\title{Extract the coefficient vector for a single decider}
\usage{
get_coefficient_vector(probit_parameters, decider_id)
}
\description{
This helper function extracts the coefficient vector with fixed and random
effects for a single decider.
}
\keyword{internal}
