% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{get_decider_identifiers}
\alias{get_decider_identifiers}
\title{Get decider identifiers}
\usage{
get_decider_identifiers(choice_identifiers)
}
\arguments{
\item{choice_identifiers}{[\code{choice_identifiers}]\cr
A \code{\link{choice_identifiers}} object.}
}
\value{
An \code{atomic} \code{vector} of decider identifiers.
}
\description{
This helper function extracts the (unique) decider identifiers from a
\code{\link{choice_identifiers}} object.
}
\keyword{internal}
