% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{get_position_from_identifier}
\alias{get_position_from_identifier}
\title{Get identifier position}
\usage{
get_position_from_identifier(
  N = length(Tp),
  Tp = 1,
  decider_number,
  occasion_number
)
}
\arguments{
\item{N}{[\code{integer(1)}]\cr
The number of deciders.}

\item{Tp}{[\code{integer(1)} | \code{integer(N)}]\cr
The number of choice occasions per decider.

Can also be of length \code{N} for a variable number of choice occasions per
decider.}

\item{decider_number}{[\code{integer(1)}]\cr
A decider number, which is a number between 1 and \code{N}.}

\item{occasion_number}{[\code{integer(1)}]\cr
An occasion number of decider \code{n}, which is a number between 1 and \code{Tp[n]}.}
}
\value{
An \code{integer} position.
}
\description{
This helper function gets a position based on a decider and a choice
occasion number.
}
\keyword{internal}
