% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_choice_set.R
\name{probit_choice_set}
\alias{probit_choice_set}
\alias{is.probit_choice_set}
\alias{print.probit_probit_choice_set}
\title{Define probit choice set}
\usage{
probit_choice_set(probit_alternatives, ranked = FALSE, delimiter = "_")

is.probit_choice_set(x)

\method{print}{probit_probit_choice_set}(x, ...)
}
\arguments{
\item{probit_alternatives}{A \code{\link{probit_alternatives}} object.}

\item{ranked}{Either \code{TRUE} for ranked choices or \code{FALSE} (default), else.}

\item{x}{A \code{\link{probit_choice_set}} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{probit_choice_set}} object. It is a \code{character}
\code{vector} with the labels for the elements in the choice set.
}
\description{
This function constructs an object of class
\code{\link{probit_choice_set}}, which defines the choice set.
}
\section{Ranked choices}{

TODO
}

