% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_choices.R
\name{probit_choices}
\alias{probit_choices}
\alias{simulate_probit_choices}
\alias{as.data.frame.probit_choices}
\alias{as.list.probit_choices}
\title{Define probit choices}
\usage{
probit_choices(
  choices,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc",
  ranked = FALSE,
  ordered = FALSE
)

simulate_probit_choices(
  probit_parameter,
  probit_covariates,
  probit_choice_set,
  seed = NULL,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc"
)

\method{as.data.frame}{probit_choices}()

\method{as.list}{probit_choices}()
}
\arguments{
\item{choices}{The observed choices in one of two possible formats:
\itemize{
\item Can be a \code{data.frame} with three columns, named according to
\code{column_choice}, \code{column_decider}, and \code{column_occasion},
where the \code{column_choice} column contains the chosen alternative
of the decider in the \code{column_decider} column at the choice occasion
in \code{column_occasion}.
\item Alternatively, it can be a \code{list}, where the \code{n}th element is a
\code{list} that has the choice of decider \code{n} at choice occasion
\code{t} as the \code{t}th element.
}}
}
\value{
A \code{\link{probit_formula}} object.
}
\description{
These functions construct and validate an object of class
\code{\link{probit_choices}}, which defines the observed choices.
}
