% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_data.R
\name{probit_data}
\alias{probit_data}
\alias{validate_probit_data}
\title{Define probit data}
\usage{
probit_data(
  probit_covariates,
  probit_choices,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc",
  ranked = FALSE,
  ordered = FALSE
)

validate_probit_data()
}
\arguments{
\item{column_choice}{TODO}

\item{column_decider}{TODO}

\item{column_occasion}{TODO}

\item{ranked}{Either \code{TRUE} for ranked choices or \code{FALSE} (default), else.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{data}{A \code{data.frame}.}
}
\value{
TODO
}
\description{
These functions construct and validate an object of class
\code{\link{probit_data}}, which defines the probit model data.
}
