% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_effects.R
\name{probit_effects}
\alias{probit_effects}
\title{Define probit model effects}
\usage{
probit_effects(probit_formula, probit_alternatives, delimiter = "_")
}
\arguments{
\item{probit_formula}{A \code{\link{probit_formula}} object.}

\item{probit_alternatives}{A \code{\link{probit_alternatives}} object.}

\item{delimiter}{A \code{character}, the delimiter between covariate and alternative name
to build the effect name.
By default, \code{delimiter = "_"}.}
}
\value{
A \code{\link{probit_effects}} object, which is a \code{data.frame}, where
each row is an effect, and columns are
1. \code{"name"}, the effect name (composed of covariate and alternative
   name),
2. \code{"covariate"}, the covariate name,
3. \code{"alternative"}, the alternative name (only if the effect is
   alternative-specific, i.e., varies across alternatives),
4. \code{"as_covariate"}, indicator whether the covariate is
   alternative-specific,
5. \code{"as_effect"}, indicator whether the effect is alternative-specific,
6. \code{"random"}, indicator whether the effect is a random effect,
7. and \code{"log_normal"}, indicator whether the random effect is
   log-normal.

The effects are ordered as follows: Fixed effects come before random effects,
and log-normal random effects are last random effects. Otherwise, the order
is determined by occurrence in \code{formula}.
}
\description{
This function constructs an object of class
\code{\link{probit_effects}}, which defines the model effects.
}
\section{Model formula}{

The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
  \item \code{choice} is the name of the dependent variable (the choices),
  \item \code{A} are names of \strong{alternative specific covariates} with
  \strong{a coefficient that is constant across alternatives},
  \item \code{B} are names of \strong{covariates that are constant across
  alternatives},
  \item and \code{C} are names of \strong{alternative specific covariates}
  with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need the special
separation form via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Random effects}{

Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (e.g., for sign-restriction) can be specified via appending
\code{+} to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\section{Base alternative}{

The full collection of coefficients for covariates that are constant across
alternatives (including alternative specific constants) is not identified.
To achieve identifiability, the coefficient of one alternative \code{base}
is typically fixed to \code{0}.
The other coefficients then have to be interpreted with respect to
\code{base}.
The base alternative is marked with a \code{*} when printing a
\code{\link{probit_alternatives}} object.
}

\examples{
probit_effects(
  probit_formula = probit_formula(
    formula = choice ~ price | income | comfort,
    re = c("price+", "income")
  ),
  probit_alternatives = probit_alternatives(J = 3)
)

}
