% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_formula.R
\name{probit_formula}
\alias{probit_formula}
\alias{is.probit_formula}
\alias{print.probit_formula}
\title{Define probit model formula}
\usage{
probit_formula(formula, re = NULL, ordered = FALSE)

is.probit_formula(x)

\method{print}{probit_formula}(x, ...)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}}, a symbolic description of the probit model,
see details.}

\item{re}{A \code{character}, the vector of names of covariates in \code{formula} that
should have a random effect, see details.
By default, \code{re = NULL}, i.e., no random effects.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{An \code{\link{probit_formula}} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{probit_formula}} object.

It contains the elements:
\describe{
\item{\code{formula}}{The model formula.}
\item{\code{re}}{The names of covariates with random effects.}
\item{\code{ordered}}{Are the choice alternatives ordered?}
\item{\code{choice}}{The name of the dependent variable.}
\item{\code{var_types}}{The three different types of covariates.}
\item{\code{ASC}}{Does the model have alternative specific constants?}
\item{\code{re_n}}{The covariates with normal mixing distribution.}
\item{\code{re_ln}}{The covariates with log-normal mixing distribution.}
}
}
\description{
This function constructs an object of class \code{\link{probit_formula}},
which defines the formula for a probit model.
}
\section{Model formula}{

The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need the special
separation form via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Random effects}{

Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (e.g., for sign-restriction) can be specified via appending
\code{+} to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

