% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitdata-package.R
\docType{package}
\name{probitdata-package}
\alias{probitdata}
\alias{probitdata-package}
\title{probitdata: Simulate Choice Data from Probit Model}
\description{
Choice data simulation from a (ranked, ordered, multinomial) probit model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/probitdata}
  \item Report bugs at \url{https://github.com/loelschlaeger/probitdata/issues}
}

}
\keyword{internal}
