% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{read_Tp}
\alias{read_Tp}
\title{Read \code{Tp}}
\usage{
read_Tp(choice_identifiers)
}
\arguments{
\item{choice_identifiers}{[\code{choice_identifiers}]\cr
A \code{\link{choice_identifiers}} object.}
}
\value{
An \code{integer} \code{vector} of length \code{N}, where \code{N} is the
number of deciders.
}
\description{
This helper function reads the number of choice occasions \code{Tp} from a
\code{\link{choice_identifiers}} object.
}
\keyword{internal}
