% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_data.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choice data}
\usage{
simulate_choices(
  RprobitB_covariates = simulate_RprobitB_covariates(formula, N, J, T = 1, alternatives =
    LETTERS[1:J], re = NULL, ordered = FALSE),
  true_parameter = RprobitB_parameter(),
  ranked = FALSE,
  seed = NULL,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc"
)
}
\arguments{
\item{RprobitB_covariates}{An \code{\link{RprobitB_covariates}} object, which contains the covariate
matrices used for the choice data simulation.}

\item{true_parameter}{An \code{\link{RprobitB_parameter}} object, which contains the model
parameters used for the choice data simulation.
By default, \code{RprobitB_parameter = RprobitB_parameter()}, i.e. default
parameters are used.}

\item{ranked}{TODO}
}
\value{
An \code{\link{RprobitB_data}} object.
}
\description{
This function simulates choice data from a probit model. It helps to create
an \code{\link{RprobitB_data}} object.
}
\examples{
### simulate data from a binary probit model with two latent classes
data <- simulate_choices(
  formula = choice ~ cost | income | time, N = 10, J = 2, T = 1:10,
  alternatives = c("car", "bus"), re = c("cost", "time"),
  true_parameter = RprobitB_parameter(C = 2)
)

### simulate data from an ordered probit model
data <- simulate_choices(
  formula = opinion ~ age + gender, N = 50, J = 5,
  alternatives = c("very bad", "bad", "indifferent", "good", "very good"),
  ordered = TRUE
)

### simulate data from a ranked probit model
data <- simulate_choices(
  formula = product ~ price, N = 10, J = 3, T = 1:10, ranked = TRUE
)

}
\seealso{
\itemize{
  \item TODO
}
}
