% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_data.R
\name{simulate_probit_data}
\alias{simulate_probit_data}
\title{Simulate choice data}
\usage{
simulate_probit_data(
  probit_covariates = sample_probit_covariates(probit_formula, N, Tp = 1,
    probit_alternatives = probit_alternatives(J = 3)),
  probit_parameter = probit_parameter(),
  ranked = FALSE,
  seed = NULL,
  column_choice = "choice"
)
}
\arguments{
\item{probit_covariates}{An \code{\link{probit_covariates}} object, which contains the covariate
matrices used for the choice data simulation.}

\item{probit_parameter}{An \code{\link{probit_parameter}} object, which contains the model
parameters used for the choice data simulation.
By default, \code{probit_parameter = probit_parameter()}, i.e. default
parameters are used.}

\item{ranked}{Either \code{TRUE} for ranked choices or \code{FALSE} (default), else.}

\item{seed}{An \code{integer}, passed to \code{set.seed()} to make the sampling of
probit covariates reproducible.
By default, \code{seed = NULL}, i.e., no seed is set.}

\item{column_choice}{TODO}
}
\value{
An \code{\link{probit_data}} object.
}
\description{
This function simulates choice data from a probit model. It helps to create
a \code{\link{probit_data}} object.
}
\section{Model formula}{

The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need the special
separation form via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Random effects}{

Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (e.g., for sign-restriction) can be specified via appending
\code{+} to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
### simulate data from a binary probit model with two latent classes
# data <- simulate_probit_data(
#   probit_covariates = sample_probit_covariates(
#     probit_formula = probit_formula(
#       formula = choice ~ cost | income | time, re = c("cost", "time")
#     ),
#     N = 10, Tp = 1:10,
#     probit_alternatives = probit_alternatives(
#       J = 2, alternatives = c("car", "bus")
#     )
#   )
# )

### simulate data from an ordered probit model
# data <- simulate_probit_data(
#   formula = opinion ~ age + gender, N = 50, J = 5,
#   alternatives = c("very bad", "bad", "indifferent", "good", "very good"),
#   ordered = TRUE
# )

### simulate data from a ranked probit model
# data <- simulate_probit_data(
#   formula = product ~ price, N = 10, J = 3, T = 1:10, ranked = TRUE
# )

}
\seealso{
\itemize{
\item TODO
}
}
