% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_parameters.R
\name{switch_parameter_space}
\alias{switch_parameter_space}
\title{TODO: inherit choice_effects and error_term}
\usage{
switch_parameter_space(choice_parameters, choice_effects, names = "effect")
}
\arguments{
\item{names}{[\code{character(1)}]\cr
\itemize{
\item \code{"effect"}
\item \code{"generic"}
\item \code{"unnamed"}
}

TODO: order of numeric vector}

\item{x}{[\code{choice_parameters} | \code{numeric}()]\cr
The choice parameters.}
}
\description{
TODO: inherit choice_effects and error_term
}
