% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_parameters.R
\name{transform_choice_parameters}
\alias{transform_choice_parameters}
\title{TODO: inherit choice_effects and error_term}
\usage{
transform_choice_parameters(x, choice_effects, mode, names = "effect")
}
\arguments{
\item{x}{[\code{choice_parameters} | \code{numeric}()]\cr
Either a \code{\link{choice_parameters}} object or a \code{numeric}
\code{vector}.}

\item{mode}{(\code{character(1)})\cr
TODO
\itemize{
\item \code{"vector"}: vector of identified parameters
\item \code{"choice_parameters"}: \code{choice_parameters} object
}}

\item{names}{(\code{character(1)})\cr
\itemize{
\item \code{"effect"}
\item \code{"generic"}
\item \code{"unnamed"}
}

TODO: order of numeric vector}
}
\description{
TODO: inherit choice_effects and error_term
}
