% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicedata-package.R
\name{change_format}
\alias{change_format}
\title{Change format of an object}
\usage{
change_format(x, new_format, ...)
}
\arguments{
\item{x}{[any]\cr
Any \code{{choicedata}} object.}

\item{new_format}{[\code{character(1)}]\cr
The requested format. Available values depend on the concrete method.}

\item{...}{Additional arguments to be passed to the corresponding methods.}
}
\value{
The transformed object.
}
\description{
This generic provides methods for changing the format of various objects
inside the \code{{choicedata}} package.
}
