% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_alternatives.R
\name{choice_alternatives}
\alias{choice_alternatives}
\alias{is.choice_alternatives}
\alias{print.choice_alternatives}
\title{Define choice alternatives}
\usage{
choice_alternatives(
  J = 2,
  alternatives = LETTERS[1:J],
  base = alternatives[1],
  ordered = FALSE
)

is.choice_alternatives(x)

\method{print}{choice_alternatives}(x, ...)
}
\arguments{
\item{J}{An \code{integer}, the number of choice alternatives.
Must be at least \code{2}.
If \code{ordered = TRUE}, must be at least \code{3}.}

\item{alternatives}{A \code{character} vector, labels for the choice alternatives.
Its length must be \code{J}.
By default, \code{alternatives = LETTERS[1:J]}.}

\item{base}{A \code{character}, the name of the base alternative for covariates that are
not alternative specific, see details.
\code{base} must be contained in \code{alternatives}.
Ignored if the model has no alternative specific covariates (e.g., in the
ordered case).
By default, \code{base} is the first element of \code{alternatives}.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{A \code{\link{choice_alternatives}} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{choice_alternatives}} object.

It contains the elements:
\describe{
\item{\code{J}}{The number of choice alternatives.}
\item{\code{alternatives}}{The labels for the choice alternatives.}
\item{\code{base}}{The name of the base alternative.}
\item{\code{ordered}}{Are the choice alternatives ordered?}
}
}
\description{
This function constructs an object of class
\code{\link{choice_alternatives}}, which defines the choice alternatives.
}
\section{Base alternative}{

The full collection of coefficients for covariates that are constant across
alternatives (including alternative specific constants) is not identified.
To achieve identifiability, the coefficient of one alternative \code{base}
is typically fixed to \code{0}.
The other coefficients then have to be interpreted with respect to
\code{base}.
The base alternative is marked with a \code{*} when printing a
\code{\link{choice_alternatives}} object.
}

