% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_covariates.R
\name{choice_covariates}
\alias{choice_covariates}
\alias{generate_choice_covariates}
\alias{as.list.choice_covariates}
\alias{as.data.frame.choice_covariates}
\alias{covariate_names}
\title{Define choice covariates}
\usage{
choice_covariates(covariates)

generate_choice_covariates(
  choice_effects,
  choice_identifiers,
  labels = covariate_names(choice_effects),
  marginals = list(),
  correlation = diag(length(labels))
)

\method{as.list}{choice_covariates}(x, ...)

\method{as.data.frame}{choice_covariates}(x, row.names, optional, ...)

covariate_names(choice_effects)
}
\arguments{
\item{covariates}{TODO}

\item{row.names, optional}{Currently not used.}
}
\value{
A \code{data.frame}.
}
\description{
The \code{choice_covariates} object defines the choice model covariates.
\itemize{
\item \code{generate_choice_covariates} samples covariates
\item \code{as.list.choice_covariates} transforms the covariates to a \code{list}
of model design matrices, see details
\item \code{as.data.frame} transforms the covariates to a \code{data.frame}
\item \code{covariate_names} provides the covariate names for a given model
specification
}
}
\section{Covariate matrices}{

A covariate matrix contains the choice covariates of a decider at some choice
occasion. It is of dimension \code{J} x \code{P}, where \code{J} is the
number of alternatives and \code{P} the number of effects. See
\code{\link{compute_P}} to compute the number \code{P}.
}

\examples{
# TODO

}
