% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_covariates.R
\name{choice_covariates}
\alias{choice_covariates}
\alias{generate_choice_covariates}
\alias{covariate_names}
\alias{change_format.choice_covariates}
\title{Define choice covariates}
\usage{
choice_covariates(
  data_frame,
  column_decider = "deciderID",
  column_occasion = "occasionID",
  as_cross_section = FALSE,
  column_covariates = setdiff(colnames(data_frame), c(column_decider, column_occasion))
)

generate_choice_covariates(
  choice_effects = NULL,
  choice_identifiers = generate_choice_identifiers(N = 100),
  labels = covariate_names(choice_effects),
  n = nrow(choice_identifiers),
  marginals = list(),
  correlation = diag(length(labels)),
  verbose = FALSE
)

covariate_names(choice_effects)

\method{change_format}{choice_covariates}(x, new_format, choice_effects)
}
\arguments{
\item{data_frame}{[\code{data.frame}]\cr
Contains the choice data.}

\item{column_decider}{[\code{character(1)}]\cr
The name of the \code{data_frame} column with identifier for the deciders.}

\item{column_occasion}{[\code{character(1)} | \code{NULL}]\cr
The name of the \code{data_frame} column with identifier for the choice
occasions.

Can also be \code{NULL} for the cross-sectional case.}

\item{as_cross_section}{[\code{logical(1)}]\cr
Treat choice data as cross-sectional?}

\item{column_covariates}{[\code{character()}]\cr
Names of columns in \code{data_frame} that store covariate values.}

\item{x}{[\code{choice_identifiers}]\cr
The \code{choice_identifiers} object to be printed.}

\item{new_format}{[\code{character(1)}]\cr
\itemize{
\item \code{"df_wide"}
\item \code{"df_long"}
\item \code{"design_matrices"}
}}
}
\value{
A \code{data.frame}.
}
\description{
The \code{choice_covariates} object defines the choice model covariates.
\itemize{
\item \code{generate_choice_covariates()} samples covariates
\item \code{design_matrices()} transforms the covariates to a \code{list}
of model design matrices, see details
\item \code{as.data.frame.design_matrices()} transforms the design matrices
(back) to a \code{data.frame}
\item \code{covariate_names()} provides the covariate names for a given model
specification
}
}
\section{Covariate matrices}{

A covariate matrix contains the choice covariates of a decider at some choice
occasion. It is of dimension \code{J} x \code{P}, where \code{J} is the
number of alternatives and \code{P} the number of effects. See
\code{\link{compute_P}} to compute the number \code{P}.
}

\examples{
# TODO

}
