% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_data.R
\name{choice_data}
\alias{choice_data}
\alias{is.choice_data}
\alias{validate_choice_data}
\title{Define choice data}
\usage{
choice_data(
  data,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc",
  ranked = FALSE,
  ordered = FALSE
)

is.choice_data(x, error = TRUE)

validate_choice_data()
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{column_choice}{A \code{character}, the column name of the \code{data.frame} with the
choices.}

\item{column_decider}{A \code{character}, the column name of the \code{data.frame} with the
decider identifiers.}

\item{column_occasion}{A \code{character}, the column name of the \code{data.frame} with the
identifiers for the choice occasions.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{An object of the corresponding class to be printed.}

\item{error}{Either \code{TRUE} to throw an error if the check failed, or \code{FALSE}
to return the check result as \code{TRUE} or \code{FALSE}, respectively.}
}
\value{
TODO
}
\description{
These functions construct and validate an object of class
\code{\link{choice_data}}, which defines the choice model data.
}
