% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_data.R
\name{choice_data}
\alias{choice_data}
\alias{is.choice_data}
\alias{validate_choice_data}
\alias{simulate_choice_data}
\alias{as.list.choice_data}
\alias{as.data.frame.choice_data}
\alias{plot.choice_data}
\alias{print.choice_data}
\title{Define choice data}
\usage{
choice_data(
  data,
  format = "wide",
  column_choice = "choice",
  column_alternative = NA,
  column_decider = "deciderID",
  column_occasion = NA,
  column_covariates_alternative_constant = character(),
  column_covariates_alternative_varying = character(),
  alternatives = character(),
  ordered = FALSE,
  ranked = FALSE,
  delimiter = "_"
)

is.choice_data(x, error = TRUE)

validate_choice_data()

simulate_choice_data(
  choice_covariates,
  choice_parameters,
  choice_formula = NULL,
  format = "wide",
  ranked = FALSE,
  column_choice = "choice"
)

\method{as.list}{choice_data}()

\method{as.data.frame}{choice_data}(x, row.names = NULL, optional = FALSE, format, ...)

\method{plot}{choice_data}(x, ...)

\method{print}{choice_data}(x, ...)
}
\arguments{
\item{format}{TODO}

\item{column_choice}{A \code{character}, the column name of the \code{data.frame} with the
choices.}

\item{column_alternative}{TODO}

\item{column_decider}{A \code{character}, the column name of the \code{data.frame} with the
decider identifiers.}

\item{column_occasion}{A \code{character}, the column name of the \code{data.frame} with the
identifiers for the choice occasions.}

\item{column_covariates_alternative_constant}{TODO}

\item{column_covariates_alternative_varying}{TODO}

\item{alternatives}{[\code{character(J)}]\cr
Labels for the choice alternatives.}

\item{ordered}{[\code{logical(1)}]\cr
Are the choice alternatives ordered?}

\item{ranked}{TODO}

\item{delimiter}{TODO}

\item{x}{[\code{choice_alternatives}]\cr
The \code{choice_alternatives} object to be printed.}

\item{error}{TODO}

\item{choice_covariates}{[\code{choice_covariates}]\cr
The choice covariates used for the choice data simulation.}

\item{choice_parameters}{An \code{\link{choice_parameters}} object, which contains the model
parameters used for the choice data simulation.}

\item{choice_formula}{TODO}

\item{row.names, optional}{Currently not used.}

\item{...}{Currently not used.}

\item{data_frame}{[\code{data.frame}]\cr
Contains the choice data.}
}
\value{
A \code{\link{choice_data}} object.
}
\description{
These functions construct and validate an object of class
\code{\link{choice_data}}, which defines the choice model data.

\code{\link{simulate_choice_data}} simulates choice data.
}
\examples{
# ### simulate data from a binary probit model with two latent classes
# data <- generate_choice_data(
#   choice_covariates = generate_choice_covariates(
#     choice_formula = choice_formula(
#       formula = choice ~ cost | income | time, re = c("cost", "time")
#     ),
#     N = 10, Tp = 1:10,
#     choice_alternatives = choice_alternatives(
#       J = 2, alternatives = c("car", "bus")
#     )
#   )
# )

### simulate data from an ordered probit model
# data <- simulate_choice_data(
#   formula = opinion ~ age + gender, N = 50, J = 5,
#   alternatives = c("very bad", "bad", "indifferent", "good", "very good"),
#   ordered = TRUE
# )

### simulate data from a ranked probit model
# data <- simulate_choice_data(
#   formula = product ~ price, N = 10, J = 3, T = 1:10, ranked = TRUE
# )

}
\seealso{
\itemize{
\item TODO
}
}
