% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_effects.R
\name{choice_effects}
\alias{choice_effects}
\alias{print.choice_effects}
\title{Define choice model effects}
\usage{
choice_effects(choice_formula, choice_alternatives, delimiter = "_")

\method{print}{choice_effects}(x, ...)
}
\arguments{
\item{choice_formula}{[\code{choice_formula}]\cr
The \code{\link{choice_formula}} object that defines the choice effects.}

\item{choice_alternatives}{[\code{choice_alternatives}]\cr
The \code{\link{choice_alternatives}} object that defines the choice effects.}

\item{delimiter}{[\code{character(1)}]\cr
A delimiter between covariate and alternative name to build the effect name.}

\item{x}{[\code{choice_effects}]\cr
The \code{choice_effects} object to be printed.}

\item{...}{Currently not used.}
}
\value{
A \code{choice_effects} object, which is a \code{data.frame}, where each row
is a model effect, and columns are
\enumerate{
\item \code{"effect_name"}, the name for the effect which is composed of
covariate and alternative name,
\item \code{"generic_name"}, the generic effect name (i.e., \code{"alpha_*"} for
non-random effects and \code{"b_*"} for random effects),
\item \code{"covariate"}, the covariate name connected to the effect,
\item \code{"alternative"}, the alternative name connected to the effect (only
if the effect is alternative-specific, i.e., varies across alternatives),
\item \code{"as_covariate"}, indicator whether the covariate is alternative-specific,
\item \code{"as_effect"}, indicator whether the effect is alternative-specific,
\item \code{"lc_effect"}, indicator whether the effect has latent classes,
\item \code{"mixing"}, a factor with levels \code{"none"}, \code{"normal"},
and \code{"log-normal"}, indicating the type of random effect.
}

For identification, the effects are ordered according to the following rules:
\itemize{
\item Non-random effects come before random effects.
\item Non-latent-class effects come before latent-class effects.
\item Normal random effects come before log-normal random effects.
\item Otherwise, the order is determined by occurrence in \code{formula}.
}

It contains the arguments \code{choice_formula}, \code{choice_alternatives}, and
\code{delimiter} as attributes.
}
\description{
This function constructs an object of class \code{choice_effects}, which
defines the effects of a choice model.
}
\examples{
choice_effects(
  choice_formula = choice_formula(
    formula = choice ~ price | income | comfort, error_term = "probit",
    random_effects = c("price+", "income"),
    latent_classes = c("comfort", "income")
  ),
  choice_alternatives = choice_alternatives(J = 3)
)

}
