% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_formula.R
\name{choice_formula}
\alias{choice_formula}
\alias{print.choice_formula}
\title{Define choice formula}
\usage{
choice_formula(formula, error_term = "probit", re = NULL)

\method{print}{choice_formula}(x, ...)
}
\arguments{
\item{formula}{(\code{formula})\cr
A symbolic description of the choice model, see details.}

\item{error_term}{(\code{character()})\cr
Defines the model's error term. Current options are:
\itemize{
\item \code{"logit"}: errors are assumed to be iid standard Gumbel distributed
\item \code{"probit"}: errors are assumed to be multivariate normally distributed
}}

\item{re}{(\code{character()})\cr
The names of covariates in \code{formula} with a random effect, see details.}

\item{x}{An object of the corresponding class.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{choice_formula}, which is a \code{list} with the
following elements:
\describe{
\item{\code{formula}}{The model formula.}
\item{\code{error_term}}{The model's error term.}
\item{\code{choice}}{The name of the discrete response variable.}
\item{\code{var_types}}{The three different types of covariates.}
\item{\code{ASC}}{Does the model have alternative-specific constants?}
\item{\code{mixing_types}}{The types of random effects (if any).}
\item{\code{ordered_valid}}{Formula valid for ordered case (see details)?}
}
}
\description{
This function constructs an object of class \code{choice_formula}, which
defines the formula for a choice model.
}
\section{Model formula}{

The structure of \code{formula} is
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the discrete response variable,
\item \code{A} are names of \strong{alternative-specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative-specific covariates}
with \strong{alternative-specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, e.g.,
\code{choice ~ A1 + A2}.

By default, alternative-specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, e.g.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second type but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
type and no ASCs.

Some parts of the formula can be omitted when there is no ambiguity. For
example, \code{choice ~ A} is equivalent to \code{choice ~ A | 1 | 0}.

In the ordered case, since only a single utility is modeled, no ASCs and no
alternative-specific covariates can be included. Hence, in the ordered case,
\code{formula} must be of the form, e.g.,  \code{choice ~ 0 | A + B + 0}.
}

\section{Random effects}{

Covariates can have random effects, i.e., their coefficients can follow a
random distribution (a so-called mixing distribution).
Per default, the mixing distribution is normal. In addition, the log-normal
distribution (e.g., for sign-restriction) can be specified via appending
\code{+} to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
choice_formula(
  formula = choice ~ A | B | C,
  error_term = "probit",
  re = c("A", "B+")
)

}
