% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_formula.R
\name{choice_formula}
\alias{choice_formula}
\alias{print.choice_formula}
\title{Define choice model formula}
\usage{
choice_formula(formula, error_term, random_effects = character())

\method{print}{choice_formula}(x, ...)
}
\arguments{
\item{formula}{[\code{formula}]\cr
A symbolic description of the choice model, see details.}

\item{error_term}{[\code{character(1)}]\cr
Defines the model's error term. Current options are:
\itemize{
\item \code{"probit"}: errors are assumed to be multivariate normally distributed
}}

\item{random_effects}{[\code{character()}]\cr
The names of covariates in \code{formula} connected to a random effect, i.e.,
their coefficients follow a random (so-called mixing) distribution.

To have random effects for the ASCs, \code{ASC} to \code{random_effects}.

Per default, the mixing distribution is normal.}

\item{x}{[\code{choice_formula}]\cr
The \code{choice_formula} object to be printed.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{choice_formula}, which is a \code{list} of the elements:
\describe{
\item{\code{formula}}{The model formula.}
\item{\code{error_term}}{The model's error term.}
\item{\code{choice}}{The name of the discrete response variable.}
\item{\code{var_types}}{The three different types of covariates.}
\item{\code{ASC}}{Does the model have alternative-specific constants?}
\item{\code{mixing_types}}{The types of random effects (if any).}
}
}
\description{
The \code{choice_formula} object defines the utility equation for a choice model.
}
\section{Choice models}{

Assume that we know the choices of \eqn{N} deciders choosing between
\eqn{J \geq 2} alternatives at each of \eqn{T} choice occasions.
Specific to each decider, alternative and choice occasion, we observe \eqn{P}
covariates, a linear combination of which explains the latent random utility:
\deqn{U_{ntj} = X_{ntj}' \tilde{\beta}_n + \epsilon_{ntj},}
\eqn{n=1,\dots,N}, \eqn{t=1,\dots,T}, and \eqn{j=1,\dots,J}.
Here, \eqn{X_{ntj}} is a (column) vector of \eqn{P} characteristics specific
to alternative \eqn{j} as faced by decider \eqn{n} at choice occasion
\eqn{t}, \eqn{\tilde{\beta}_n \in \mathbb{R}^{P}} is the coefficient vector
encoding the preferences of \eqn{n}, and
\eqn{(\epsilon_{nt:}) = (\epsilon_{nt1},\dots,\epsilon_{ntJ})'} is the
model's error term vector for \eqn{n} at \eqn{t}.

In the probit model case, the error vector \eqn{(\epsilon_{nt:})} is normally
distributed with covariance matrix \code{Sigma}.

The value \eqn{U_{ntj}} can be interpreted as the decider's utility for
alternative \eqn{j}. It is unobserved by the researcher, but we assume that
the deciders know their utilities for each alternative and make a choice
which is consistent with utility maximization. Therefore,
\deqn{y_{nt} = \operatorname*{argmax}_{j = 1,\dots,J} U_{ntj},}
where \eqn{y_{nt}=j} denotes the event that decider \eqn{n} chooses \eqn{j}
at their \eqn{t}-th choice occasion.

Entries of the decider-specific coefficient vector \eqn{\tilde{\beta}_n} can
be fixed across deciders, in which case the coefficient vector is of the form
\eqn{\tilde{\beta}_n' = (\alpha', \beta_n')'}, where
\eqn{\alpha \in \mathbb{R}^{P_f}} are \eqn{P_f} coefficients that are
constant across deciders and \eqn{\beta_n} are \eqn{P_r} decider-specific
coefficients, \eqn{P_f + P_r = P}.

The decider-specific coefficients are assumed to be random effects, i.e.,
realizations of an underlying mixing distribution and to be independent of
the characteristics \eqn{X_{ntj}} and the errors \eqn{(\epsilon_{nt:})}.
This distribution characterizes heterogeneity among the deciders and allows
for individual sensitivities.
}

\section{Specifying the model formula}{

The structure of \code{formula} is \code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the discrete response variable,
\item \code{A} are names of \strong{alternative-specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative-specific covariates}
with \strong{alternative-specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, e.g.,
\code{choice ~ A1 + A2}.

By default, alternative-specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, e.g.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second type but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
type and no ASCs.

Some parts of the formula can be omitted when there is no ambiguity. For
example, \code{choice ~ A} is equivalent to \code{choice ~ A | 1 | 0}.

In the ordered case, since only a single utility is modeled, no ASCs and no
alternative-specific covariates can be included. Hence, in the ordered case,
\code{formula} must be of the form, e.g., \code{choice ~ 0 | A + B + 0}.
}

\examples{
choice_formula(
  formula = choice ~ A | B | C,
  error_term = "probit",
  random_effects = c("A", "B")
)

}
