% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{choice_identifiers}
\alias{choice_identifiers}
\alias{print.choice_identifiers}
\alias{generate_choice_identifiers}
\title{Define choice identifiers}
\usage{
choice_identifiers(
  data_frame,
  column_decider = "deciderID",
  column_occasion = "occasionID",
  as_cross_section = FALSE
)

\method{print}{choice_identifiers}(x, ...)

generate_choice_identifiers(
  N,
  Tp = 1,
  column_decider = "deciderID",
  column_occasion = "occasionID"
)
}
\arguments{
\item{data_frame}{[\code{data.frame}]\cr
Contains the choice data.}

\item{column_decider}{[\code{character(1)}]\cr
The name of the \code{data.frame} column with identifier for the deciders.}

\item{column_occasion}{[\code{character(1)} | \code{NULL}]\cr
The name of the \code{data.frame} column with identifier for the choice
occasions.

Can also be \code{NULL} in the cross-sectional case.}

\item{as_cross_section}{[\code{logical(1)}]\cr
Treat choice data as cross-sectional?}

\item{x}{[\code{choice_identifiers}]\cr
The \code{choice_identifiers} object to be printed.}

\item{...}{Currently not used.}

\item{N}{[\code{integer(1)}]\cr
The number of deciders.}

\item{Tp}{[\code{integer(1)} | \code{integer(N)}]\cr
The constant number of choice occasions per decider.

Can also be of length \code{N} for a variable number of choice occasions per
decider.}
}
\value{
An object of class \code{choice_identifiers} object, which is a
\code{data.frame} with two columns:
\enumerate{
\item \code{column_decider} contains the decider ids,
\item \code{column_occasion} contains the choice occasion ids.
}
}
\description{
The \code{choice_identifiers} object defines identifiers for the deciders and
the choice occasions.
}
\examples{
generate_choice_identifiers(N = 2, Tp = 2)

}
