% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{choice_identifiers}
\alias{choice_identifiers}
\alias{is.choice_identifiers}
\alias{generate_choice_identifiers}
\alias{read_choice_identifiers}
\title{Define identifiers}
\usage{
choice_identifiers(
  data_frame = data.frame(),
  column_decider = "deciderID",
  column_occasion = NULL,
  as_cs = FALSE
)

is.choice_identifiers(x, error = TRUE)

generate_choice_identifiers(
  N,
  Tp = 1,
  column_decider = "deciderID",
  column_occasion = NULL
)

read_choice_identifiers(
  data_frame = data.frame(),
  column_decider = "deciderID",
  column_occasion = NULL,
  as_cs = FALSE
)
}
\arguments{
\item{data_frame}{A \code{data.frame}.}

\item{column_decider}{A \code{character}, the name of the \code{data.frame} column with identifier
for the deciders. The default is \code{"deciderID"}.}

\item{column_occasion}{A \code{character}, different from \code{column_decider}, the name of the
\code{data.frame} column with identifier for the choice occasions. These
identifiers must be unique for a given decider.

It can also be \code{NULL} (default), in which case data is treated as
cross-sectional. The decider ids must be unique in this case. A column for
the (trivial) choice occasion identifiers is generated anyways with the
default name \code{"occasionID"}.}

\item{as_cs}{Either \code{TRUE} to treat the data as cross-sectional (i.e., dropping the
information of multiple choices by the same decider), or \code{FALSE}
(default), else.}

\item{N}{A positive \code{integer}, the number of deciders.}

\item{Tp}{A positive \code{integer}, the number of choice occasions per decider.
Can also be a \code{vector} of length \code{N} for a variable number of
choice occasions per decider.
By default, \code{Tp = 1}.}
}
\value{
A \code{\link{choice_identifiers}} object. It is a \code{data.frame} with
two columns, one named \code{column_decider} and containing the decider ids,
and the other one either named \code{column_occasion} or \code{"occasionID"} by
default, containing the choice occasion ids.
}
\description{
These functions define an object of class \code{\link{choice_identifiers}},
which defines identifiers for the deciders and the choice occasions.

\code{\link{generate_choice_identifiers}} generates choice identifiers.

\code{\link{read_choice_identifiers}} reads choice identifiers from a
\code{data.frame}.
}
