% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_parameters.R
\name{choice_parameters}
\alias{choice_parameters}
\alias{is.choice_parameters}
\alias{sample_choice_parameters}
\alias{validate_choice_parameters}
\alias{print.choice_parameters}
\alias{as.vector.choice_parameters}
\title{Define choice model parameters}
\usage{
choice_parameters(
  C = 1,
  s = NA,
  alpha = NA,
  b = NA,
  Omega = NA,
  Sigma = NA,
  Sigma_diff = NA,
  diff_alt = 1,
  beta = NA,
  z = NA,
  d = NA,
  model_type = NA_character_,
  ordered = NA
)

is.choice_parameters(x)

sample_choice_parameters(
  x = choice_parameters(),
  model_type = "probit",
  formula,
  re = NULL,
  ordered = FALSE,
  J,
  N,
  seed = NULL,
  validate = TRUE
)

validate_choice_parameters(
  x = choice_parameters(),
  formula,
  re = NULL,
  ordered = FALSE,
  J,
  N
)

\method{print}{choice_parameters}(
  x,
  ...,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = FALSE,
  details = !simplify
)

\method{as.vector}{choice_parameters}(x, mode = "numeric")
}
\arguments{
\item{C}{An \code{integer}, the number (greater or equal 1) of latent classes of
decision makers.
By default, \code{C = 1}.}

\item{s}{A \code{numeric} vector of length \code{C}, the vector of class weights.
For identifiability, the vector elements must be decreasing.}

\item{alpha}{A \code{numeric} vector of length \code{P_f}, the vector of fixed effect
coefficients.
If \code{C > 1}, a \code{matrix} of dimension \code{P_f} x \code{C},
where column \code{c} contains the fixed effect coefficients for class
\code{c}.}

\item{b}{A \code{numeric} vector of length \code{P_r}, the vector of mean random
effects.
If \code{C > 1}, a \code{matrix} of dimension \code{P_r} x \code{C},
where column \code{c} contains the mean random effects for class
\code{c}.}

\item{Omega}{A \code{matrix} of dimension \code{P_r} x \code{P_r}, the covariance matrix
of random effects.
If \code{C > 1}, a \code{matrix} of dimension \code{P_r^2} x \code{C},
where column \code{c} contains the covariance matrix of random effects for
class \code{c} in vector form.}

\item{Sigma}{A \code{matrix} of dimension \code{J} x \code{J}, the error term covariance
matrix for a probit model.
In the ordered probit model (see details), \code{Sigma} is a
\code{matrix} of dimension \code{1} x \code{1} (or simply a single
\code{numeric}).}

\item{Sigma_diff}{A \code{matrix} of dimension \code{J-1} x \code{J-1}, the differenced error
term covariance matrix for a probit model.
\code{Sigma_diff} is assumed to be differenced with respect to alternative
\code{diff_alt}, see details.
\code{Sigma_diff} is ignored in case of the ordered probit model
(see details) or if \code{Sigma} is specified.}

\item{diff_alt}{An \code{integer} from \code{1} to \code{J}, the reference alternative for
utility differencing that maps \code{Sigma} to \code{Sigma_diff}, see
details.
By default, \code{diff_alt = 1}.}

\item{beta}{A \code{matrix} of dimension \code{P_r} x \code{N}, the matrix of the
decider-specific coefficient vectors.
The coefficient vector for decider \code{n} is stored in column \code{n}.}

\item{z}{A \code{numeric} vector of length \code{N}, the vector of the allocation
variables.
Entry \code{n} of \code{z} is an integer from \code{1} to \code{C} and
denotes the allocated class for decider \code{n}.}

\item{d}{A \code{numeric} of length \code{J - 2}, the vector of logarithmic increases
of the utility thresholds.
Only relevant in the ordered model case (see details).}

\item{model_type}{Either \code{"probit"} (default) to select the probit model, or
\code{"logit"} to select the logit model.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{A \code{\link{choice_parameters}} object.}

\item{formula}{A \code{\link[stats]{formula}}, a symbolic description of the choice model,
see details.}

\item{re}{A \code{character}, the vector of names of covariates in \code{formula} that
should have a random effect, see details.
By default, \code{re = NULL}, i.e., no random effects.}

\item{J}{An \code{integer}, the number of choice alternatives.
Must be at least \code{2}.
If \code{ordered = TRUE}, must be at least \code{3}.}

\item{N}{A positive \code{integer}, the number of deciders.}

\item{seed}{An \code{integer}, passed to \code{set.seed()} to make the sampling of
missing choice parameters reproducible.
By default, \code{seed = NULL}, i.e., no seed is set.}

\item{...}{A \code{character} (vector), the names of model parameters to be printed.
By default, all available parameters are printed.}

\item{rowdots}{An \code{integer}, the row number which is replaced by \code{...}.
By default, \code{rowdots = 4}.}

\item{coldots}{An \code{integer}, the column number which is replaced by \code{...}.
By default, \code{coldots = 4}.}

\item{digits}{An \code{integer}, the number of printed decimal places.
Only relevant if input \code{x} is numeric.
By default, \code{digits = 2}.}

\item{simplify}{A \code{logical}, set to \code{TRUE} to simplify the output.
By default, \code{simplify = FALSE}.}

\item{details}{A \code{logical}, set to \code{TRUE} to print the type and
dimension of \code{x}.
By default, \code{details = !simplify}.}
}
\value{
A \code{\link{choice_parameters}} object.

It contains the elements:
\describe{
\item{\code{C}}{The number of latent classes.}
\item{\code{s}}{The class weights.}
\item{\code{alpha}}{The fixed coefficients.}
\item{\code{b}}{The class means.}
\item{\code{Omega}}{The class covariances.}
\item{\code{Sigma}}{The error term covariance matrix.}
\item{\code{Sigma_diff}}{The differenced error term covariance matrix.}
\item{\code{diff_alt}}{The reference alternative for utility differencing.}
\item{\code{beta}}{The decider-specific coefficient vectors.}
\item{\code{z}}{The class allocation variables.}
\item{\code{d}}{The logarithmic increases of the utility thresholds.}
}
}
\description{
These functions construct and validate an object of class
\code{\link{choice_parameters}}, which defines the parameters of a choice
model, see details.

\code{sample_choice_parameters()} draws (missing) choice model
parameters at random, see details.
}
\section{Setting choice model parameters}{

\enumerate{
\item Use \code{choice_parameters()} to construct a
\code{\link{choice_parameters}} object, where any model parameter can be
specified (see below).
\item Next, call \code{validate_choice_parameters()} with the
\code{\link{choice_parameters}} object created in step 1. This will add
unspecified parameters (see below for details) and validate all specified
parameters.
}
}

\section{The probit model}{

Assume that we know the choices of \eqn{N} deciders choosing between
\eqn{J \geq 2} alternatives at each of \eqn{T} choice occasions.
Specific to each decider, alternative and choice occasion, we observe \eqn{P}
covariates, a linear combination of which eventually explains the latent
random utility:
\deqn{U_{ntj} = X_{ntj}' \tilde{\beta}_n + \epsilon_{ntj},}
\eqn{n=1,\dots,N}, \eqn{t=1,\dots,T}, and \eqn{j=1,\dots,J}.
Here, \eqn{X_{ntj}} is a (column) vector of \eqn{P} characteristics specific
to alternative \eqn{j} as faced by decider \eqn{n} at choice occasion
\eqn{t}, \eqn{\tilde{\beta}_n \in \mathbb{R}^{P}} is the coefficient vector
encoding the preferences of \eqn{n}, and
\eqn{(\epsilon_{nt:}) = (\epsilon_{nt1},\dots,\epsilon_{ntJ})' \sim
\text{MVN}_{J} (0,\Sigma)} is the model's error term vector for \eqn{n} at
\eqn{t}.

The value \eqn{U_{ntj}} can be interpreted as the decider's utility for
alternative \code{j}.
It is unobserved by the researcher, but we assume that the deciders know
their utilities for each alternative and make a choice which is consistent
with utility maximization. Therefore,
\deqn{y_{nt} = \operatorname*{argmax}_{j = 1,\dots,J} U_{ntj},}
where \eqn{y_{nt}=j} denotes the event that decider \eqn{n} chooses \eqn{j}
at her \eqn{t}-th choice occasion.

Entries of the decider-specific coefficient vector \eqn{\tilde{\beta}_n} can
be fixed across deciders, in which case the coefficient vector is of the form
\eqn{\tilde{\beta}_n' = (\alpha', \beta_n')'}, where
\eqn{\alpha \in \mathbb{R}^{P_f}} are \eqn{P_f} coefficients that are
constant across deciders and \eqn{\beta_n} are \eqn{P_r} decider-specific
coefficients, \eqn{P_f + P_r = P}.

The decider-specific coefficients are assumed to be realizations of an
underlying mixing distribution and to be independent of the characteristics
\eqn{X_{ntj}} and the errors \eqn{(\epsilon_{nt:})}.
This distribution characterizes heterogeneity among the deciders and allows
for individual sensitivities. As mixing distribution, we assume a mixture of
\eqn{P_r}-variate Gaussian densities \eqn{\phi_{P_r}} with mean vectors
\eqn{b = (b_c)_{c}} and covariance matrices \eqn{\Omega = (\Omega_c)_{c}}
using \eqn{C} components:
\deqn{\beta_n\mid b,\Omega \sim \sum_{c=1}^{C} s_c \phi_{P_r} (\cdot \mid
b_c,\Omega_c).}
Here, \eqn{(s_c)_{c}} are weights satisfying \eqn{0 < s_c\leq 1} for
\eqn{c=1,\dots,C} and \eqn{\sum_c s_c=1}.

One interpretation of the latent class model is obtained by introducing
variables \eqn{z=(z_n)_n}, allocating each decision maker \eqn{n} to class
\eqn{c} with probability \eqn{s_c}, i.e.,
\deqn{\text{Prob}(z_n=c)=s_c \land \beta_n \mid z,b,\Omega \sim
\phi_{P_r}(\cdot \mid b_{z_n},\Omega_{z_n}).}
}

\section{Ordered probit model}{

When the set of choice alternatives is ordered, the probit model has only a
single utility
\deqn{U_{nt} = X_{nt}' \tilde{\beta}_n + \epsilon_{nt},}
\eqn{\epsilon_{nt} \sim \text{MVN}_{1} (0,\Sigma)},
per decider \eqn{n} and choice occasion \eqn{t}. The utility can be
interpreted as the level of association that \eqn{n} has with the choice
question. It falls into discrete categories, which in turn are linked to the
ordered alternatives \eqn{j=1,\dots,J}. Formally,
\deqn{y_{nt} = \sum_{j = 1,\dots,J} j \cdot I(\gamma_{j-1} < U_{nt} \leq
\gamma_{j}),}
where \eqn{\gamma_0 = -\infty} and \eqn{\gamma_J = +\infty}. This implies
that alternative \eqn{j} is chosen, if the utility falls into the interval
\eqn{(\gamma_{j-1}, \gamma_j]}.
Monotonicity of the thresholds \eqn{(\gamma_j)_{j=1,\dots,J-1}} is ensured
by estimating logarithmic increments \eqn{d_j} with
\eqn{\gamma_j = \sum_{i\leq j} \exp{(d_i)}}, \eqn{j=1,\dots,J-1}.
For level normalization, we fix \eqn{\gamma_1 = 0}.
}

\section{Level and scale normalization}{

Choice models are invariant towards the level and scale of utility, hence
a transformation is required for identifiability.

For level normalization, we take utility differences:
\deqn{\tilde{U}_{ntj} = \tilde{X}_{ntj}' \tilde{\beta}_n +
\tilde{\epsilon}_{ntj},}
where (choosing some alternative \eqn{k \in \{1,\dots,J\}} as the reference,
also denoted by \code{diff_alt})
\eqn{\tilde{U}_{ntj} = U_{ntj} - U_{ntk}},
\eqn{\tilde{X}_{ntj} = X_{ntj} - X_{ntk}}, and
\eqn{\tilde{\epsilon}_{ntj} = \epsilon_{ntj} - \epsilon_{ntk}} for
\eqn{j\neq k}.
The error term differences \eqn{(\tilde{\epsilon}_{nt:})} again are
multivariate normally distributed with mean \eqn{0} but transformed
covariance matrix \eqn{\tilde{\Sigma}}, also denoted by \code{Sigma_diff}.
See \code{\link[oeli]{diff_cov}} for computing \code{Sigma_diff} from
\code{Sigma}, and \code{\link[oeli]{undiff_cov}} for the other way around.

For level normalization in the ordered model case, we fix
\eqn{\gamma_1 = 0}.

For scale normalization, we fix the top left element of \code{Sigma_diff} to
\eqn{1} (or fix \code{Sigma = 1} in the ordered probit case).
}

\section{Drawing missing choice model parameters}{


Unspecified choice model parameters are drawn independently from the
following distributions:
\describe{
\item{\code{s}}{The class weights are drawn from a Dirichlet distribution
with concentration 1.}
\item{\code{alpha}}{The fixed coefficients are drawn from a
multivariate normal distribution with zero mean and a diagonal covariance
matrix with value 10 on the diagonal.}
\item{\code{b}}{The class means are drawn from a
multivariate normal distribution with zero mean and a diagonal covariance
matrix with value 10 on the diagonal.}
\item{\code{Omega}}{The class covariances are drawn from an Inverse-Wishart
distribution with degrees of freedom equal to \code{P_r} + 2 and scale
matrix equal to the identity matrix.}
\item{\code{Sigma}}{The error term covariance matrix is drawn from an
Inverse-Wishart distribution with degrees of freedom equal to \code{J} + 2
and scale matrix equal to the identity matrix.}
\item{\code{Sigma_diff}}{The differenced error term covariance matrix is
derived from \code{Sigma}.}
\item{\code{beta}}{The decider-specific coefficient vectors are drawn from
a multivariate normal distribution with mean defined by \code{b} and
covariance defined by \code{Omega}.
}
\item{\code{z}}{The class allocation variables are drawn from a discrete
distribution over the class labels with probabilities defined by \code{s}.}
\item{\code{d}}{The logarithmic increases of the utility thresholds are
drawn from a multivariate normal distribution with zero mean and a diagonal
covariance matrix with value 10 on the diagonal.}
}
}

\examples{
(x <- choice_parameters(C = 2))
formula <- choice ~ A | B
re <- "A"
J <- 3
N <- 100
(x <- validate_choice_parameters(x, formula = formula, re = re, J = J, N = N))

}
