% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_parameters.R
\name{choice_parameters}
\alias{choice_parameters}
\alias{print.choice_parameters}
\alias{generate_choice_parameters}
\alias{validate_choice_parameters}
\alias{identify_choice_parameters}
\title{Define choice model parameters}
\usage{
choice_parameters(alpha = NULL, b = NULL, Omega = NULL, Sigma = NULL)

\method{print}{choice_parameters}(
  x,
  ...,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = FALSE,
  details = !simplify
)

generate_choice_parameters(
  choice_effects,
  fixed_parameters = choice_parameters()
)

validate_choice_parameters(
  choice_parameters,
  choice_effects,
  allow_missing = FALSE
)

identify_choice_parameters(choice_parameters, scale = 1)
}
\arguments{
\item{alpha}{[\code{numeric(P_f)}]\cr
The non-random coefficients. Only relevant if non-random coefficients are
specified.}

\item{b}{[\code{numeric(P_r)}]\cr
The mean of random effects. Only relevant if random coefficients are
specified.}

\item{Omega}{[`matrix(P_r)``]\cr
The covariance matrix of random effects. Only relevant if random coefficients
are specified.}

\item{Sigma}{[\code{matrix(nrow = J, ncol = J)}]\cr
The error term covariance matrix. Only relevant in the probit model.}

\item{x}{[\code{choice_parameters}]\cr
A \code{choice_parameters} object.}

\item{...}{Currently not used.}

\item{rowdots}{[\code{integer(1)}]\cr
The row number which is replaced by \code{...}.}

\item{coldots}{[\code{integer(1)}]\cr
The column number which is replaced by \code{...}.}

\item{digits}{[\code{integer(1)}]\cr
The number of printed decimal places if input \code{x} is \code{numeric}.}

\item{simplify}{[\code{logical(1)}]\cr
Simplify the output?}

\item{details}{[\code{logical(1)}]\cr
Print the type and dimension of \code{x}?}

\item{choice_effects}{[\code{choice_effects}]\cr
The \code{\link{choice_effects}} object that defines the choice effects.}

\item{fixed_parameters}{[\code{choice_parameters}]\cr
Optionally a \code{choice_parameters} object of parameters to keep
fixed when sampling other parameters.}

\item{choice_parameters}{[\code{choice_parameters}]\cr
A \code{choice_parameters} object.}

\item{scale}{TODO}
}
\value{
An object of class \code{choice_parameters}, which is a \code{list} with the elements:
\describe{
\item{\code{alpha}}{The non-random coefficients (if any).}
\item{\code{b}}{The mean of random effects (if any).}
\item{\code{Omega}}{The covariance of random effects (if any).}
\item{\code{Sigma}}{The error term covariance (if any).}
}
}
\description{
These functions construct, validate, and transform an object of class
\code{choice_parameters}, which defines the parameters of a choice model.
\itemize{
\item \code{choice_parameters()} constructs a \code{choice_parameters} object.
\item \code{generate_choice_parameters()} samples parameters at random, see details.
\item \code{validate_choice_parameters()} validates a \code{choice_parameters} object.
\item \code{identify_choice_parameters()} applies scale and level normalization, see
details.
\item \code{change_format.choice_parameters()} transforms between the interpretation
and optimization parameter space, see details.
}
}
\section{Sampling missing choice model parameters}{


Unspecified choice model parameters (if required for the model) are drawn
independently from the following distributions:
\describe{
\item{\code{alpha}}{Drawn from a multivariate normal distribution with
zero mean and a diagonal covariance matrix with value 10 on the diagonal.}
\item{\code{b}}{Drawn from a multivariate normal distribution with zero
mean and a diagonal covariance matrix with value 10 on the diagonal.}
\item{\code{Omega}}{Drawn from an Inverse-Wishart distribution with degrees
of freedom equal to \code{P_r} + 2 and scale matrix equal to the identity.}
\item{\code{Sigma}}{Drawn from an Inverse-Wishart distribution with degrees
of freedom equal to \code{J} + 2 and scale matrix equal to the identity.}
}
}

\section{Level and scale normalization}{

Choice models are invariant towards the level and scale of utility, hence
a transformation is required for identifiability.

For level normalization, we replace the first row and the first column of
\code{Sigma} with zeros, which corresponds to taking utility differences with
respect to alternative \code{1}.

For scale normalization, we fix the \eqn{(2, 2)}-value of \code{Sigma} to a
positive value \code{scale}.
}

\section{Choice models}{

Assume that we know the choices of \eqn{N} deciders choosing between
\eqn{J \geq 2} alternatives at each of \eqn{T} choice occasions.
Specific to each decider, alternative and choice occasion, we observe \eqn{P}
covariates, a linear combination of which explains the latent random utility:
\deqn{U_{ntj} = X_{ntj}' \tilde{\beta}_n + \epsilon_{ntj},}
\eqn{n=1,\dots,N}, \eqn{t=1,\dots,T}, and \eqn{j=1,\dots,J}.
Here, \eqn{X_{ntj}} is a (column) vector of \eqn{P} characteristics specific
to alternative \eqn{j} as faced by decider \eqn{n} at choice occasion
\eqn{t}, \eqn{\tilde{\beta}_n \in \mathbb{R}^{P}} is the coefficient vector
encoding the preferences of \eqn{n}, and
\eqn{(\epsilon_{nt:}) = (\epsilon_{nt1},\dots,\epsilon_{ntJ})'} is the
model's error term vector for \eqn{n} at \eqn{t}.

In the probit model case, the error vector \eqn{(\epsilon_{nt:})} is normally
distributed with covariance matrix \code{Sigma}.

The value \eqn{U_{ntj}} can be interpreted as the decider's utility for
alternative \eqn{j}. It is unobserved by the researcher, but we assume that
the deciders know their utilities for each alternative and make a choice
which is consistent with utility maximization. Therefore,
\deqn{y_{nt} = \operatorname*{argmax}_{j = 1,\dots,J} U_{ntj},}
where \eqn{y_{nt}=j} denotes the event that decider \eqn{n} chooses \eqn{j}
at their \eqn{t}-th choice occasion.

Entries of the decider-specific coefficient vector \eqn{\tilde{\beta}_n} can
be fixed across deciders, in which case the coefficient vector is of the form
\eqn{\tilde{\beta}_n' = (\alpha', \beta_n')'}, where
\eqn{\alpha \in \mathbb{R}^{P_f}} are \eqn{P_f} coefficients that are
constant across deciders and \eqn{\beta_n} are \eqn{P_r} decider-specific
coefficients, \eqn{P_f + P_r = P}.

The decider-specific coefficients are assumed to be random effects, i.e.,
realizations of an underlying mixing distribution and to be independent of
the characteristics \eqn{X_{ntj}} and the errors \eqn{(\epsilon_{nt:})}.
This distribution characterizes heterogeneity among the deciders and allows
for individual sensitivities.
}

\examples{
### generate choice parameters at random
choice_effects <- choice_effects(
  choice_formula = choice_formula(
    formula = choice ~ A | B, error_term = "probit", random_effects = "A"
  ),
  choice_alternatives = choice_alternatives(J = 3)
)
choice_parameters <- generate_choice_parameters(
  choice_effects = choice_effects,
  fixed_parameters = choice_parameters("b" = 1)
)

}
