% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_parameters.R
\name{choice_parameters}
\alias{choice_parameters}
\alias{print.choice_parameters}
\alias{generate_choice_parameters}
\alias{validate_choice_parameters}
\alias{identify_choice_parameters}
\title{Define choice model parameters}
\usage{
choice_parameters(
  s = NULL,
  alpha = NULL,
  b = NULL,
  Omega = NULL,
  Sigma = NULL,
  gamma = NULL
)

\method{print}{choice_parameters}(
  x,
  ...,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = FALSE,
  details = !simplify
)

generate_choice_parameters(
  choice_effects,
  C = 1,
  fixed_parameters = choice_parameters()
)

validate_choice_parameters(
  choice_parameters,
  choice_effects,
  C = 1,
  allow_missing = FALSE
)

identify_choice_parameters(choice_parameters, scale = 1, level = 1)
}
\arguments{
\item{s}{[\code{numeric(C)}]\cr
The latent class weights. Only relevant if latent classes are specified.

Must be non-negative and sum-up to 1.}

\item{alpha}{[\code{numeric(P_f)} | \code{matrix(nrow = P_f, ncol = C)}]\cr
The non-random coefficients. Only relevant if non-random coefficients are
specified.

In the case of latent classes, different values are stored column-wise.}

\item{b}{[\code{numeric(P_r)} | \code{matrix(nrow = P_r, ncol = C)}]\cr
The mean of random effects. Only relevant if random coefficients are
specified.

In the case of latent classes, different values are stored column-wise.}

\item{Omega}{[\code{matrix(P_r)} | \code{matrix(nrow = P_r^2, ncol = C)}]\cr
The covariance matrix of random effects. Only relevant if random coefficients
are specified.

In the case of latent classes, different covariance matrices are stored
column-wise in vector form.}

\item{Sigma}{[\code{matrix(nrow = J, ncol = J)} | \code{numeric(1)}]\cr
The error term covariance matrix. Only relevant in the probit model.

In the ordered probit model, \code{Sigma} is \code{numeric(1)}.}

\item{gamma}{[\code{numeric(J - 1)}]\cr
The utility thresholds. Only relevant in the ordered model case.
Must be strictly ascending.

\code{gamma} corresponds to \eqn{\gamma_1, \dots, \gamma_{J-1}}, while the
lower and upper bounds are \eqn{\gamma_0 = -\infty} and
\eqn{\gamma_J = +\infty}.}

\item{x}{[\code{choice_parameters}]\cr
The \code{choice_parameters} object to be printed.}

\item{...}{Currently not used.}

\item{rowdots}{[\code{integer(1)}]\cr
The row number which is replaced by \code{...}.}

\item{coldots}{[\code{integer(1)}]\cr
The column number which is replaced by \code{...}.}

\item{digits}{[\code{integer(1)}]\cr
The number of printed decimal places if input \code{x} is \code{numeric}.}

\item{simplify}{[\code{logical(1)}]\cr
Simplify the output?}

\item{details}{[\code{logical(1)}]\cr
Print the type and dimension of \code{x}?}

\item{choice_effects}{[\code{choice_effects}]\cr
The \code{\link{choice_effects}} object that defines the choice effects.}

\item{C}{[\code{integer(1)}]\cr
The number of latent classes (if any specified in \code{choice_formula}).

By default, \code{C = 1}, which effectively means no latent classes.}

\item{fixed_parameters}{[\code{choice_parameters}]\cr
Optionally a \code{choice_parameters} object of parameters to keep
fixed when sampling other parameters.}

\item{choice_parameters}{TODO}

\item{scale}{TODO}

\item{level}{TODO}
}
\value{
An object of class \code{choice_parameters}, which is a \code{list} with the elements:
\describe{
\item{\code{s}}{The latent class weights (if any).}
\item{\code{alpha}}{The non-random coefficients (if any).}
\item{\code{b}}{The mean of random effects (if any).}
\item{\code{Omega}}{The covariance of random effects (if any).}
\item{\code{Sigma}}{The error term covariance (if any).}
\item{\code{gamma}}{The utility thresholds (if any).}
}
}
\description{
These functions construct, validate, and transform an object of class
\code{choice_parameters}, which defines the parameters of a choice model.
\itemize{
\item \code{choice_parameters()} constructs a \code{choice_parameters} object.
\item \code{generate_choice_parameters()} samples parameters at random, see details.
\item \code{validate_choice_parameters()} validates a \code{choice_parameters} object.
\item \code{identify_choice_parameters()} applies scale and level normalization.
\item \code{change_format.choice_parameters()} transforms between parameter spaces.
}
}
\section{Sampling missing choice model parameters}{


Unspecified choice model parameters (if required for the model) are drawn
independently from the following distributions:
\describe{
\item{\code{s}}{drawn from a Dirichlet distribution with concentration 1}
\item{\code{alpha}}{drawn from a multivariate normal distribution with
zero mean and a diagonal covariance matrix with value 10 on the diagonal}
\item{\code{b}}{drawn from a multivariate normal distribution with zero
mean and a diagonal covariance matrix with value 10 on the diagonal}
\item{\code{Omega}}{drawn from an Inverse-Wishart distribution with degrees
of freedom equal to \code{P_r} + 2 and scale matrix equal to the identity}
\item{\code{Sigma}}{drawn from an Inverse-Wishart distribution with degrees
of freedom equal to \code{J} + 2 and scale matrix equal to the identity,
in the ordered probit case drawn from a standard normal distribution}
\item{\code{gamma}}{derived from the logarithmic increases of the utility
thresholds \code{d}, which are drawn from a multivariate normal
distribution with zero mean and covariance matrix equal to the identity}
}
}

\section{Level and scale normalization}{

Choice models are invariant towards the level and scale of utility, hence
a transformation is required for identifiability.

For level normalization, we take utility differences:
\deqn{\tilde{U}_{ntj} = \tilde{X}_{ntj}' \tilde{\beta}_n +
\tilde{\epsilon}_{ntj},}
where (choosing some alternative \eqn{k \in \{1,\dots,J\}} as the reference)
\eqn{\tilde{U}_{ntj} = U_{ntj} - U_{ntk}},
\eqn{\tilde{X}_{ntj} = X_{ntj} - X_{ntk}}, and
\eqn{\tilde{\epsilon}_{ntj} = \epsilon_{ntj} - \epsilon_{ntk}} for
\eqn{j\neq k}.

In the probit model case, the error term differences
\eqn{(\tilde{\epsilon}_{nt:})} again are
multivariate normally distributed with mean \eqn{0} but transformed
covariance matrix \eqn{\tilde{\Sigma}}, also denoted by \code{Sigma_diff}.
See \code{\link[oeli]{diff_cov}} for computing \code{Sigma_diff} from
\code{Sigma}, and \code{\link[oeli]{undiff_cov}} for the other way around.

For level normalization in the ordered model case, we fix
\eqn{\gamma_1 = 0}.

For scale normalization, we fix the top left element of \code{Sigma_diff} to
the value \code{scale} (or fix \code{Sigma = scale} in the ordered probit case).
In the logit model case, the scale normalization is already implied by the
assumption that the errors are independently extreme value distributed with
variance \eqn{\pi^2/6}.
}

\section{Choice models}{

Assume that we know the choices of \eqn{N} deciders choosing between
\eqn{J \geq 2} alternatives at each of \eqn{T} choice occasions.
Specific to each decider, alternative and choice occasion, we observe \eqn{P}
covariates, a linear combination of which explains the latent random utility:
\deqn{U_{ntj} = X_{ntj}' \tilde{\beta}_n + \epsilon_{ntj},}
\eqn{n=1,\dots,N}, \eqn{t=1,\dots,T}, and \eqn{j=1,\dots,J}.
Here, \eqn{X_{ntj}} is a (column) vector of \eqn{P} characteristics specific
to alternative \eqn{j} as faced by decider \eqn{n} at choice occasion
\eqn{t}, \eqn{\tilde{\beta}_n \in \mathbb{R}^{P}} is the coefficient vector
encoding the preferences of \eqn{n}, and
\eqn{(\epsilon_{nt:}) = (\epsilon_{nt1},\dots,\epsilon_{ntJ})'} is the
model's error term vector for \eqn{n} at \eqn{t}.

In the probit model case, the error vector \eqn{(\epsilon_{nt:})} is normally
distributed with covariance matrix \code{Sigma}.

The value \eqn{U_{ntj}} can be interpreted as the decider's utility for
alternative \eqn{j}. It is unobserved by the researcher, but we assume that
the deciders know their utilities for each alternative and make a choice
which is consistent with utility maximization. Therefore,
\deqn{y_{nt} = \operatorname*{argmax}_{j = 1,\dots,J} U_{ntj},}
where \eqn{y_{nt}=j} denotes the event that decider \eqn{n} chooses \eqn{j}
at their \eqn{t}-th choice occasion.

Entries of the decider-specific coefficient vector \eqn{\tilde{\beta}_n} can
be fixed across deciders, in which case the coefficient vector is of the form
\eqn{\tilde{\beta}_n' = (\alpha', \beta_n')'}, where
\eqn{\alpha \in \mathbb{R}^{P_f}} are \eqn{P_f} coefficients that are
constant across deciders and \eqn{\beta_n} are \eqn{P_r} decider-specific
coefficients, \eqn{P_f + P_r = P}.

The decider-specific coefficients are assumed to be random effects, i.e.,
realizations of an underlying mixing distribution and to be independent of
the characteristics \eqn{X_{ntj}} and the errors \eqn{(\epsilon_{nt:})}.
This distribution characterizes heterogeneity among the deciders and allows
for individual sensitivities.
}

\examples{
choice_formula <- choice_formula(
  formula = choice ~ A | B, error_term = "probit", random_effects = "A"
)
choice_alternatives <- choice_alternatives(J = 3)
generate_choice_parameters(
  choice_effects = choice_effects(choice_formula, choice_alternatives)
)

}
