% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicedata-package.R
\docType{package}
\name{choicedata-package}
\alias{choicedata}
\alias{choicedata-package}
\title{choicedata: Working with Choice Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Offers a set of objects tailored to simplify working with choice data. It enables the computation of choice probabilities and the likelihood of various types of choice models based on given data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/choicedata}
  \item \url{https://loelschlaeger.de/choicedata/}
  \item Report bugs at \url{https://github.com/loelschlaeger/choicedata/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

Authors:
\itemize{
  \item Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})
}

}
\keyword{internal}
