% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicedata-package.R
\docType{package}
\name{choicedata-package}
\alias{choicedata}
\alias{choicedata-package}
\title{choicedata: Simulate, Read, and Visualize Choice Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a set of objects and methods designed to facilitate the simulation, reading and visualization of choice data generated by various types of probit and logit models (including binary, multinomial, ordered, ranked, mixed, and latent class models) and to compute the likelihood of these models for given data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/choicedata}
  \item Report bugs at \url{https://github.com/loelschlaeger/choicedata/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

}
\keyword{internal}
