% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{choices}
\alias{choices}
\alias{simulate_choices}
\alias{is.choices}
\alias{validate_choices}
\alias{as.data.frame.choices}
\alias{as.list.choices}
\title{Define choices}
\usage{
choices(
  choices = list(),
  column_choice = column_choice,
  column_decider = "deciderID",
  column_occasion = "occasionID",
  ranked = FALSE,
  ordered = FALSE,
  delimiter = "_"
)

simulate_choices(
  choice_parameters,
  choice_covariates,
  choice_formula = NULL,
  ordered = FALSE,
  ranked = FALSE
)

is.choices(x, error = TRUE)

validate_choices()

\method{as.data.frame}{choices}(x, ...)

\method{as.list}{choices}(x, ...)
}
\arguments{
\item{choices}{A \code{list}, where the \code{n}-th element is a \code{list} of choices for
the \code{n}-th decider, where the \code{t}-th element is an element from
\code{choice_set} that defines the choice at their \code{t}-th choice
occasion.}

\item{column_choice}{A \code{character}, the column name of the \code{data.frame} with the
choices.}

\item{column_decider}{A \code{character}, the column name of the \code{data.frame} with the
decider identifiers.}

\item{column_occasion}{A \code{character}, the column name of the \code{data.frame} with the
identifiers for the choice occasions.}
}
\value{
A \code{\link{choices}} object. It can be in \code{data.frame} or \code{list}
and can be transformed between these two formats via the functions
\code{as.data.frame.choices()} or \code{as.list.choices()}.
\itemize{
\item \code{data.frame} format: a \code{data.frame} with three columns
named \code{column_choice} (the choices), \code{column_decider}
(the decider identifiers), and \code{column_occasion} (the
identifiers for the choice occasions)
\item \code{list} format: a \code{list}, where the \code{n}-th element is
a \code{list} of choices for the \code{n}-th decider, where the
\code{t}-th element is an element from \code{choice_set} that defines
the choice at their \code{t}-th choice occasion
}
}
\description{
These functions construct and validate an object of class
\code{\link{choices}}, which defines the observed choices.
}
\section{Ranked choices}{

Ranked choices are yet another model variation: rather than recording only
the single most preferred alternative, some surveys ask for a full ranking of
all the alternatives, which reveals far more about the underlying
preferences. Ranked choices can by analyzed by setting \code{ranked = TRUE}.
The choice column of the data set must provide the full ranking for each
choice occasion (from most preferred to least preferred), where the
alternatives are separated by the \code{delimiter} string.

The ranked probit model follows directly from the general unordered case
noting that the ranking implies that the highest ranked alternative is chosen
in any case, while the second highest ranked alternative is chosen, if the
highest ranked alternative is not available and so forth. The only difference
is that we take flexible utility differences such that the differenced
utility vector is always negative, in contrast to the general case where we
difference with respect to a fixed reference alternative. Thereby, we
incorporate information of the full ranking.
}

