% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{choices}
\alias{choices}
\alias{simulate_choices}
\alias{is.choices}
\alias{validate_choices}
\alias{as.data.frame.choices}
\alias{as.list.choices}
\title{Define choices}
\usage{
choices(
  choices = list(),
  column_choice = column_choice,
  column_decider = "id",
  column_occasion = "idc",
  ranked = FALSE,
  ordered = FALSE,
  delimiter = "_"
)

simulate_choices(
  choice_parameters,
  choice_covariates,
  choice_set,
  seed = NULL,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc"
)

is.choices(x)

validate_choices()

\method{as.data.frame}{choices}(x, ...)

\method{as.list}{choices}(x, ...)
}
\arguments{
\item{choices}{A \code{list}, where the \code{n}-th element is a \code{list} of choices for
the \code{n}-th decider, where the \code{t}-th element is an element from
\code{choice_set} that defines the choice at their \code{t}-th choice
occasion.}

\item{column_choice}{A \code{character}, the column name of the \code{data.frame} with the
choices.}

\item{column_decider}{A \code{character}, the column name of the \code{data.frame} with the
decider identifiers.}

\item{column_occasion}{A \code{character}, the column name of the \code{data.frame} with the
identifiers for the choice occasions.}

\item{ranked}{Either \code{TRUE} for ranked choices or \code{FALSE} (default), else.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{choice_parameters}{A \code{\link{choice_parameters}} object.}

\item{choice_covariates}{A \code{\link{choice_covariates}} object.}

\item{choice_set}{A \code{\link{choice_set}} object.}

\item{seed}{An \code{integer}, passed to \code{set.seed()} to make the simulation of
choices reproducible.
By default, \code{seed = NULL}, i.e., no seed is set.}

\item{x}{An \code{\link{choices}} object.}
}
\value{
A \code{\link{choices}} object. It can be in \code{data.frame} or \code{list}
and can be transformed between these two formats via the functions
\code{as.data.frame.choices()} or \code{as.list.choices()}.
\itemize{
\item \code{data.frame} format: a \code{data.frame} with three columns
named \code{column_choice} (the choices), \code{column_decider}
(the decider identifiers), and \code{column_occasion} (the
identifiers for the choice occasions)
\item \code{list} format: a \code{list}, where the \code{n}-th element is
a \code{list} of choices for the \code{n}-th decider, where the
\code{t}-th element is an element from \code{choice_set} that defines
the choice at their \code{t}-th choice occasion
}
}
\description{
These functions construct and validate an object of class
\code{\link{choices}}, which defines the observed choices.
}
