% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_effects.R
\name{compute_P}
\alias{compute_P}
\title{Count model effects}
\usage{
compute_P(choice_effects)
}
\arguments{
\item{choice_effects}{[\code{choice_effects}]\cr
A \code{\link{choice_effects}} object.}
}
\value{
An integer with the number of model effects.
}
\description{
\code{compute_P()} returns the total number of effects in a choice model.
}
\keyword{model}
