% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_effects.R
\name{compute_P}
\alias{compute_P}
\alias{compute_P_f}
\alias{compute_P_r}
\title{Compute number of model effects}
\usage{
compute_P(formula, random_effects, J, ordered = FALSE)

compute_P_f(formula, random_effects, J, ordered = FALSE)

compute_P_r(formula, random_effects, J, ordered = FALSE)
}
\arguments{
\item{formula}{(\code{formula})\cr
A symbolic description of the choice model, see details.}

\item{random_effects}{(\code{character()})\cr
The names of covariates in \code{formula} connected to a random effect, i.e.,
their coefficients follow a random (so-called mixing) distribution.

Per default, the mixing distribution is normal. In addition, the log-normal
distribution (e.g., for sign-restriction) can be specified via appending
\code{+} to the corresponding name. To have random effects for the ASCs, add
\code{ASC} (or \code{ASC+}) to \code{random_effects}.}

\item{J}{(\code{integer(1)})\cr
The number of choice alternatives.

Must be at least \code{2}.}

\item{ordered}{(\code{logical(1)})\cr
Are the choice alternatives ordered?}
}
\value{
An \code{integer}, the number of model effects.
}
\description{
These helper functions compute the number effects in a specified choice
model:
\itemize{
\item \code{compute_P()} computes the total number \code{P} of model effects.
\item \code{compute_P_f()} computes the number \code{P_f} of non-random
model effects.
\item \code{compute_P_r()} computes the number \code{P_r} of random model
effects.
}
}
\section{Specifying the model formula}{

The structure of \code{formula} is
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the discrete response variable,
\item \code{A} are names of \strong{alternative-specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative-specific covariates}
with \strong{alternative-specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, e.g.,
\code{choice ~ A1 + A2}.

By default, alternative-specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, e.g.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second type but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
type and no ASCs.

Some parts of the formula can be omitted when there is no ambiguity. For
example, \code{choice ~ A} is equivalent to \code{choice ~ A | 1 | 0}.

In the ordered case, since only a single utility is modeled, no ASCs and no
alternative-specific covariates can be included. Hence, in the ordered case,
\code{formula} must be of the form, e.g.,  \code{choice ~ 0 | A + B + 0}.
}

\section{Ordered choice alternatives}{

When the set of choice alternatives is ordered, the choice model has only a
single utility equation
\deqn{U_{nt} = X_{nt}' \tilde{\beta}_n + \epsilon_{nt},}
where \eqn{\epsilon_{nt} \sim \text{MVN}_{1} (0,\Sigma)} in the probit model
and logistic in the logit model, per decider \eqn{n} and choice
occasion \eqn{t}.

This utility can be interpreted as the level of association that \eqn{n} has
with the choice question. It falls into discrete categories, which in turn
are linked to the ordered alternatives \eqn{j=1,\dots,J}. Formally,
\deqn{y_{nt} = \sum_{j = 1,\dots,J} j \cdot I(\gamma_{j-1} < U_{nt} \leq
\gamma_{j}),}
where \eqn{\gamma_0 = -\infty} and \eqn{\gamma_J = +\infty}. This implies
that alternative \eqn{j} is chosen, if the utility falls into the interval
\eqn{(\gamma_{j-1}, \gamma_j]}.
Monotonicity of the thresholds \eqn{(\gamma_j)_{j=1,\dots,J-1}} is ensured
by estimating logarithmic increments \eqn{d_j} with
\eqn{\gamma_j = \sum_{i\leq j} \exp{(d_i)}}, \eqn{j=1,\dots,J-1}.
For level normalization, we fix \eqn{\gamma_1 = 0}.
}

\keyword{internal}
