% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_effects.R
\name{compute_P}
\alias{compute_P}
\alias{compute_P_f}
\alias{compute_P_r}
\title{Number of model effects}
\usage{
compute_P(choice_effects)

compute_P_f(choice_effects)

compute_P_r(choice_effects)
}
\arguments{
\item{choice_effects}{[\code{choice_effects}]\cr
The \code{\link{choice_effects}} object that defines the choice effects.}
}
\value{
An \code{integer}, the number of model effects.
}
\description{
These helper functions count the number of model effects:
\itemize{
\item \code{compute_P()} returns the total number \code{P} of model effects.
\item \code{compute_P_f()} returns the number \code{P_f} of non-random effects.
\item \code{compute_P_r()} returns the number \code{P_r} of random effects.
}
}
\section{Choice models}{

Assume that we know the choices of \eqn{N} deciders choosing between
\eqn{J \geq 2} alternatives at each of \eqn{T} choice occasions.
Specific to each decider, alternative and choice occasion, we observe \eqn{P}
covariates, a linear combination of which explains the latent random utility:
\deqn{U_{ntj} = X_{ntj}' \tilde{\beta}_n + \epsilon_{ntj},}
\eqn{n=1,\dots,N}, \eqn{t=1,\dots,T}, and \eqn{j=1,\dots,J}.
Here, \eqn{X_{ntj}} is a (column) vector of \eqn{P} characteristics specific
to alternative \eqn{j} as faced by decider \eqn{n} at choice occasion
\eqn{t}, \eqn{\tilde{\beta}_n \in \mathbb{R}^{P}} is the coefficient vector
encoding the preferences of \eqn{n}, and
\eqn{(\epsilon_{nt:}) = (\epsilon_{nt1},\dots,\epsilon_{ntJ})'} is the
model's error term vector for \eqn{n} at \eqn{t}.

In the probit model case, the error vector \eqn{(\epsilon_{nt:})} is normally
distributed with covariance matrix \code{Sigma}.

The value \eqn{U_{ntj}} can be interpreted as the decider's utility for
alternative \eqn{j}. It is unobserved by the researcher, but we assume that
the deciders know their utilities for each alternative and make a choice
which is consistent with utility maximization. Therefore,
\deqn{y_{nt} = \operatorname*{argmax}_{j = 1,\dots,J} U_{ntj},}
where \eqn{y_{nt}=j} denotes the event that decider \eqn{n} chooses \eqn{j}
at their \eqn{t}-th choice occasion.

Entries of the decider-specific coefficient vector \eqn{\tilde{\beta}_n} can
be fixed across deciders, in which case the coefficient vector is of the form
\eqn{\tilde{\beta}_n' = (\alpha', \beta_n')'}, where
\eqn{\alpha \in \mathbb{R}^{P_f}} are \eqn{P_f} coefficients that are
constant across deciders and \eqn{\beta_n} are \eqn{P_r} decider-specific
coefficients, \eqn{P_f + P_r = P}.

The decider-specific coefficients are assumed to be random effects, i.e.,
realizations of an underlying mixing distribution and to be independent of
the characteristics \eqn{X_{ntj}} and the errors \eqn{(\epsilon_{nt:})}.
This distribution characterizes heterogeneity among the deciders and allows
for individual sensitivities.
}

