% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_effects.R
\name{compute_P}
\alias{compute_P}
\alias{compute_P_f}
\alias{compute_P_r}
\title{Compute number of model effects}
\usage{
compute_P(formula, re, J, ordered = FALSE)

compute_P_f(formula, re, J, ordered = FALSE)

compute_P_r(formula, re, J, ordered = FALSE)
}
\arguments{
\item{formula}{(\code{formula})\cr
A symbolic description of the choice model, see details.}

\item{re}{(\code{character()})\cr
The names of covariates in \code{formula} with a random effect, see details.}

\item{J}{(\code{integer(1)})\cr
The number of choice alternatives.

Must be at least \code{2}.}

\item{ordered}{(\code{logical(1)})\cr
Are the choice alternatives ordered?}
}
\value{
An \code{integer}, the number of model effects.
}
\description{
These functions compute the number of model effects.
\itemize{
\item \code{compute_P()} computes the total number \code{P} of model effects.
\item \code{compute_P_f()} computes the number \code{P_f} of non-random
model effects.
\item \code{compute_P_r()} computes the number \code{P_r} of random model
effects.
}
}
\section{Model formula}{

The structure of \code{formula} is
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the discrete response variable,
\item \code{A} are names of \strong{alternative-specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative-specific covariates}
with \strong{alternative-specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, e.g.,
\code{choice ~ A1 + A2}.

By default, alternative-specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, e.g.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second type but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
type and no ASCs.

Some parts of the formula can be omitted when there is no ambiguity. For
example, \code{choice ~ A} is equivalent to \code{choice ~ A | 1 | 0}.

In the ordered case, since only a single utility is modeled, no ASCs and no
alternative-specific covariates can be included. Hence, in the ordered case,
\code{formula} must be of the form, e.g.,  \code{choice ~ 0 | A + B + 0}.
}

\section{Random effects}{

Covariates can have random effects, i.e., their coefficients can follow a
random distribution (a so-called mixing distribution).
Per default, the mixing distribution is normal. In addition, the log-normal
distribution (e.g., for sign-restriction) can be specified via appending
\code{+} to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\keyword{internal}
