% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_identifiers.R
\name{decider_identifiers_to_cross_section}
\alias{decider_identifiers_to_cross_section}
\title{Unique decider identifiers in cross-sectional case}
\usage{
decider_identifiers_to_cross_section(
  decider_identifiers,
  occasion_identifiers,
  delimiter = "."
)
}
\arguments{
\item{decider_identifiers, occasion_identifiers}{[\code{atomic()}]\cr
An \code{atomic} \code{vector} of identifiers.}
}
\value{
An \code{atomic} \code{vector} of unique identifiers.
}
\description{
This helper function makes unique decider identifiers for a given combination
of decider and occasion identifiers.
}
\keyword{internal}
