% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicedata-helpers.R
\name{doc-helper}
\alias{doc-helper}
\title{Documentation helper}
\arguments{
\item{choice_alternatives}{A \code{\link{choice_alternatives}} object.}

\item{choice_covariates}{A \code{\link{choice_covariates}} object.}

\item{choice_formula}{A \code{\link{choice_formula}} object.}

\item{choice_parameters}{A \code{\link{choice_parameters}} object.}

\item{ranked}{Either \code{TRUE} for ranked choices or \code{FALSE} (default), else.}

\item{x}{An object of the corresponding class.}

\item{...}{Currently not used.}
}
\description{
Documentation helper
}
\keyword{internal}
