% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_covariates.R
\name{expand_Tp}
\alias{expand_Tp}
\title{Expand \code{Tp}}
\usage{
expand_Tp(N, Tp = 1)
}
\arguments{
\item{N}{A positive \code{integer}, the number of deciders.}

\item{Tp}{A positive \code{integer}, the number of choice occasions per decider.
Can also be a \code{vector} of length \code{N} for a variable number of
choice occasions per decider.
By default, \code{Tp = 1}.}
}
\value{
An \code{integer} \code{vector} of length \code{N}.
}
\description{
This function expands the number of choice occasions \code{Tp} to a
\code{vector} of length \code{N}.
}
\keyword{internal}
