% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_covariates.R
\name{probit_covariates}
\alias{probit_covariates}
\alias{is.probit_covariates}
\alias{sample_probit_covariates}
\alias{validate_probit_covariates}
\alias{as.list.probit_covariates}
\alias{as.data.frame.probit_covariates}
\alias{covariate_names}
\alias{covariate_number}
\title{Define choice covariates}
\usage{
probit_covariates(probit_data)

is.probit_covariates(x)

sample_probit_covariates(
  probit_formula,
  N,
  Tp = 1,
  probit_alternatives,
  seed = NULL,
  covariate_levels = Inf,
  occasion_constant = character(),
  covariate_mean = 0,
  covariate_sd = 1,
  covariate_correlation = 0,
  delimiter = "_",
  column_decider = "id",
  column_occasion = "idc"
)

validate_probit_covariates(
  covariates,
  N,
  Tp = 1,
  probit_formula,
  probit_alternatives,
  delimiter,
  column_decider,
  column_occasion
)

\method{as.list}{probit_covariates}(x)

\method{as.data.frame}{probit_covariates}(x)

covariate_names(probit_formula, probit_alternatives, delimiter = "_")

covariate_number(probit_formula, probit_alternatives)
}
\arguments{
\item{probit_data}{A \code{\link{probit_data}} object.}

\item{x}{An \code{\link{probit_covariates}} object.}

\item{probit_formula}{A \code{\link{probit_formula}} object.}

\item{N}{A positive \code{integer}, the number of deciders.}

\item{Tp}{A positive \code{integer}, the number of choice occasions per decider.
Can also be a \code{vector} of length \code{N} for a variable number of
choice occasions per decider.
By default, \code{Tp = 1}.}

\item{probit_alternatives}{A \code{\link{probit_alternatives}} object.}

\item{covariate_levels}{Defines the number of levels for the covariates. Two formats are allowed:
\itemize{
\item Can be a single, positive \code{numeric} which defines the number of levels
for all covariates. By default, \code{covariate_levels = Inf}.
\item Can be a named \code{numeric} \code{vector} for specific mean values for
specific covariates. Names must correspond to the output of
\code{\link{covariate_names}}. Means for missing covariates are set to
\code{0}.
}}

\item{occasion_constant}{A \code{character} \code{vector} of covariate names
(see \code{\link{covariate_names}})
that are supposed to be constant across choice occasions.}

\item{covariate_mean}{Defines the mean for the covariates. Two formats are allowed:
\itemize{
\item Can be a single \code{numeric} which defines a mean for all covariates.
By default, \code{covariate_mean = 0}.
\item Can be a named \code{numeric} \code{vector} for specific mean values for
specific covariates. Names must correspond to the output of
\code{\link{covariate_names}}. Means for missing covariates are set to
\code{0}.
}}

\item{covariate_sd}{Defines the standard deviation for the covariates. Two formats are allowed:
\itemize{
\item Can be a single, positive \code{numeric} which defines a standard deviation
for all covariates. By default, \code{covariate_sd = 0}.
\item Can be a named \code{numeric} \code{vector} for specific standard
deviations for specific covariates. Names must correspond to the output of
\code{\link{covariate_names}}. Standard deviations for missing covariates
are set to \code{1}.
}}

\item{covariate_correlation}{Defines the correlation across covariates. Two formats are allowed:
\itemize{
\item Can be a single \code{numeric} between \code{-1} and \code{1} which defines
the correlation across all covariates.
By default, \code{covariate_correlation = 0}.
\item Can be a correlation \code{matrix} of dimension \code{n}, where \code{n}
must equal the output of \code{\link{covariate_number}}. The
\code{(i,j)}-th element defines the correlation between covariate \code{i}
and covariate \code{j} in the output of \code{\link{covariate_names}}.
}}
}
\value{
An \code{\link{probit_covariates}} object. It is a \code{list} of
\code{list}s of \code{matrix} elements.

More precise: Let the return value be \code{out}, then
\itemize{
\item \code{out} contains the covariate matrices of all deciders at all choice
occasions,
\item \code{out[[n]]} is a \code{list} of the covariate matrices of decider
\code{n} at all of their choice occasions,
\item \code{out[[n]][[t]]} is the covariate \code{matrix} of decider \code{n} at
their \code{t}-th choice occasion.
}
}
\description{
These functions create and validate an object of class
\code{\link{probit_covariates}}, which contains the model covariate,
see the details.

\code{\link{sample_probit_covariates}} samples covariates.

\code{\link{as.list.probit_covariates}} and
\code{\link{as.data.frame.probit_covariates}} transform the covariates in
\code{list} or \code{data.frame} form, respectively.

\code{\link{covariate_names}} and \code{\link{covariate_numbers}} provide the
names and number of covariates for a given model specification.
}
\section{Covariate matrices}{

A covariate matrix contains the choice covariates of a decider at some choice
occasion. It is of dimension \code{J} x \code{P}, where \code{J} is the
number of alternatives and \code{P} the number of effects. See
\code{\link{compute_P}} to compute the number \code{P}.

By default, covariates are sampled from independent normal distributions, but
customization is possible:
\itemize{
\item the levels per covariate can be defined via \code{covariate_levels},
\item covariates with constant value over occasions can be defined via
\code{occasion_constant},
\item the means for each covariate can be defined via \code{covariate_mean},
\item the standard deviation for each covariate can be defined via
\code{covariate_sd},
\item the correlation between covariates can be defined via
\code{covariate_correlation}.
}
}

\keyword{object}
