% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit_data.R
\name{probit_data}
\alias{probit_data}
\title{Define probit data}
\usage{
probit_data(
  data,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc",
  ranked = FALSE,
  ordered = FALSE
)
}
\arguments{
\item{data}{TBA}
}
\description{
These functions construct and validate an object of class
\code{\link{probit_data}}, which defines the probit model data.
}
