% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitdata-package.R
\docType{package}
\name{probitdata-package}
\alias{probitdata}
\alias{probitdata-package}
\title{probitdata: Manage Probit Choice Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a set of functions designed to facilitate the simulation, reading and visualization of choice data generated by various types of probit models (including ranked, ordered, multinomial, and latent class models) and to compute the likelihood of these models for a given data set.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/probitdata}
  \item Report bugs at \url{https://github.com/loelschlaeger/probitdata/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

}
\keyword{internal}
