% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probitdata-package.R
\docType{package}
\name{probitdata-package}
\alias{probitdata}
\alias{probitdata-package}
\title{probitdata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/probitdata}
  \item Report bugs at \url{https://github.com/loelschlaeger/probitdata/issues}
}

}
\keyword{internal}
