% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{student_smoking_choice}
\alias{student_smoking_choice}
\title{Ordered Smoking Responses of Students}
\format{
A \code{tibble} with 237 rows and 4 columns:
\describe{
\item{student [\code{integer}]}{The student identifier.}
\item{choice [\code{ordered}]}{The smoking level.}
\item{age [\code{numeric}]}{The student's age in years.}
\item{exercise [\code{factor}]}{Whether the student exercises frequently,
sometimes, or not at all.}
}
}
\source{
Adapted from \code{survey} in the
\href{https://CRAN.R-project.org/package=MASS}{\pkg{MASS} package}.
The original student survey is documented by Venables and Ripley (2002).
}
\usage{
student_smoking_choice
}
\description{
Ordered, cross-sectional smoking responses from 237 Statistics I students
at the University of Adelaide. The response levels are never, occasional,
regular, and heavy.

This data can be used to model the choice of smoking based on age and
exercise frequency.
}
\references{
\insertRef{VenablesRipley2002}{choicedata}
}
\keyword{dataset}
