% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{travel_mode_choice}
\alias{travel_mode_choice}
\title{Revealed Preferences for Travel Mode Choice}
\format{
A \code{tibble} with 840 rows and 8 columns:
\describe{
\item{individual [\code{integer}]}{The identifier for the decider.}
\item{mode [\code{character}]}{The travel mode.}
\item{choice [\code{integer}]}{Whether the mode was chosen.}
\item{wait [\code{integer}]}{The terminal waiting time, 0 for car.}
\item{cost [\code{integer}]}{The travel cost in dollars.}
\item{travel [\code{integer}]}{The travel time in minutes.}
\item{income [\code{integer}]}{The household income in thousand dollars.}
\item{size [\code{integer}]}{The traveling group size.}
}
}
\usage{
travel_mode_choice
}
\description{
Data set of revealed choices by 210 travelers between Sydney and Melbourne
who report their choice between the four travel modes plane, train, bus, or
car. The data were collected as part of a 1987 intercity mode choice study.
}
\references{
\insertRef{BenAkiva1993}{choicedata}
}
\keyword{dataset}
